% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotextgraph.R
\name{biotextgraph}
\alias{biotextgraph}
\title{biotextgraph}
\usage{
biotextgraph(target, argList)
}
\arguments{
\item{target}{"pubmed", "bugsigdb", "refseq", "ec"}

\item{argList}{passed to each function}
}
\value{
list of data frames and ggplot2 object
}
\description{
wrapper for functions wcGeneSummary, wcAbst, and wcBSDB
}
\examples{
geneList <- c("DDX41","PNKP")
biotextgraph("refseq", argList=list(geneList=geneList))
}
