% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{changeLayout}
\alias{changeLayout}
\title{changeLayout}
\usage{
changeLayout(g, layout_func)
}
\arguments{
\item{g}{osplot object}

\item{layout_func}{layout function in igraph}
}
\description{
changeLayout
}
