% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{connectGenes}
\alias{connectGenes}
\title{connectGenes}
\usage{
connectGenes(nets, query_word)
}
\arguments{
\item{nets}{named list of nets}

\item{query_word}{word to connect}
}
\description{
When an interesting word is found in two or more networks,
connect the words and gene names and return the graph.
Note that genePlot must be set to TRUE.
}
