% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convertMetaCyc}
\alias{convertMetaCyc}
\title{convertMetaCyc}
\usage{
convertMetaCyc(ids, onlySpecies = FALSE)
}
\arguments{
\item{ids}{tax ids from metacyc}

\item{onlySpecies}{parse only species}
}
\value{
coverted species name
}
\description{
Convert MetaCyc IDs to the actual taxonomy names.
}
\details{
This function needs taxonomizr package to be installed,
and needs download of databases by prepareDatabase(getAccessions=FALSE)
}
\examples{
\dontrun{convertMetaCyc("TAX-9606")}
}
