% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enzyme.R
\name{enzyme}
\alias{enzyme}
\title{enzyme}
\usage{
enzyme(
  file,
  ecnum,
  onlyTerm = FALSE,
  onlyDf = FALSE,
  target = "abstract",
  taxec = FALSE,
  taxFile = NULL,
  candTax = NULL,
  argList = list(),
  apiKey = NULL
)
}
\arguments{
\item{file}{file downloaded from expasy}

\item{ecnum}{candidate ecnum, like those obtained from eggNOG-mapper}

\item{onlyTerm}{only return quoted queries to wcAbst}

\item{onlyDf}{only return ec description data.frame
if onlyTerm and onlyDf are both specified, onlyTerm have priority}

\item{target}{abstract or title}

\item{taxec}{link taxonomy to EC using UniProt Taxonomy ID file
If this is TRUE, data.frame is returned}

\item{taxFile}{UniProt organism ID file path}

\item{candTax}{when taxec=TRUE, search only for these species.}

\item{argList}{passed to osplot(target="pubmed")}

\item{apiKey}{api key for PubMed}
}
\value{
object consisting of data frame and ggplot2 object
}
\description{
Text mining the enzyme information.
}
\details{
Query the Enzyme Comission number and obtain description,
and search pubmed for these enzymes and make word cloud and
correlation network. Need to specify the path to "enzyme.dat"
downloaded from from ExPASy (https://enzyme.expasy.org/).
}
\examples{
file <- "enzyme.dat"
\dontrun{enzyme(file, ecnum="1.2.1.1")}
}
\seealso{
generalf
}
