% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exportCyjs}
\alias{exportCyjs}
\title{exportCyjs}
\usage{
exportCyjs(g, rootDir, netDir, bubble = FALSE, withEdge = TRUE)
}
\arguments{
\item{g}{igraph object}

\item{rootDir}{root directory path}

\item{netDir}{directory to store scripts}

\item{bubble}{default to FALSE, if node attribute "group" is available,
use the grouping to draw bubblesets using
`https://github.com/upsetjs/cytoscape.js-bubblesets`}

\item{withEdge}{connect edges or not in bubblesets}
}
\value{
return nothing, export to a specified directory
}
\description{
Export Cytoscape.js from the network
}
\details{
Export Cytoscape.js script, HTML and stylesheet for the graph and image
}
\examples{
library(igraph)
g <- graph_from_literal( ME1-+ME2 )
V(g)$image <- c("path1","path2")
V(g)$shape <- c("image","image")
V(g)$size <- c(1,1)
\dontrun{exportCyjs(g, "./", "net")}
}
