% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exportCyjsWithoutImage}
\alias{exportCyjsWithoutImage}
\title{exportCyjsWithoutImage}
\usage{
exportCyjsWithoutImage(
  g,
  rootDir,
  netDir,
  edgeWidth = "weight",
  nodeColor = "tag",
  nodeSize = "Freq",
  nodeColorDiscretePal = "RdBu",
  sizeMin = 10,
  sizeMax = 50
)
}
\arguments{
\item{g}{igraph object}

\item{rootDir}{root directory path}

\item{netDir}{directory to store scripts}

\item{edgeWidth}{attribute name for edgeWidth}

\item{nodeColor}{attribute name for node color}

\item{nodeSize}{attribute name for node size}

\item{nodeColorDiscretePal}{color mapping palette for discrete variables}

\item{sizeMin}{minimum size for scaling node size}

\item{sizeMax}{maximum size for scaling node size}
}
\value{
return nothing, export to a specified directory
}
\description{
Export Cytoscape.js from the network without images
}
\details{
Export Cytoscape.js script, HTML and stylesheet for the graph without image
}
\examples{
library(igraph)
g <- graph_from_literal( ME1-+ME2 )
V(g)$size <- c(1,1)
\dontrun{exportCyjsWithoutImage(g, "./", "net")}
}
