% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exportVisjs}
\alias{exportVisjs}
\title{exportVisjs}
\usage{
exportVisjs(g, rootDir, netDir)
}
\arguments{
\item{g}{igraph object}

\item{rootDir}{root directory path}

\item{netDir}{directory to store scripts}
}
\description{
Export vis.js script, HTML and stylesheet for the graph and image
}
