% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportWCNetwork.R
\name{exportWCNetwork}
\alias{exportWCNetwork}
\title{exportWCNetwork}
\usage{
exportWCNetwork(
  g,
  geneList,
  dir = "network",
  colors = NULL,
  sizeMin = 50,
  sizeMax = 200,
  wcScale = 20,
  keyType = "SYMBOL",
  orgDb = org.Hs.eg.db,
  wcArgs = list(),
  numWords = 50
)
}
\arguments{
\item{g}{igraph (node name corresponding to gene cluster name)}

\item{geneList}{named list of gene IDs}

\item{dir}{directory to output the files, default to current directory}

\item{colors}{color palettes to use for each cluster
named list, or random palette will be assigned if NULL}

\item{sizeMin}{minimum scale size for nodes}

\item{sizeMax}{maximum scale size for nodes}

\item{wcScale}{scaling size for wordcloud}

\item{keyType}{type of gene IDs (default to SYMBOL)}

\item{orgDb}{annotation database}

\item{wcArgs}{parameters for ggwordcloud}

\item{numWords}{number of words}
}
\value{
export the Cytoscape.js network
}
\description{
Export wordcloud networks to Cytoscape.js
}
\details{
The function exports the wordcloud network by providing the gene cluster network.
Node size will be the number of genes in the cluster.
By default use cola layout, and see below for the parameters of the layout.
(https://github.com/cytoscape/cytoscape.js-cola) 
If "strength" attribute is in the edge of igraph object,
the parameter is used to size the edge width.
}
\examples{
g <- igraph::graph_from_edgelist(rbind(c("ME1","ME2")))
gl <- list("ME1"=c("PNKP","IRF3","MCM5"), "ME2"=c("IRF3","ERCC2","TP53"))
exportWCNetwork(g, gl)
}
