% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alliance.R, R/bugsigdb.R, R/manual.R,
%   R/pubmed.R, R/refseq.R
\name{alliance}
\alias{alliance}
\alias{bugsigdb}
\alias{manual}
\alias{pubmed}
\alias{generalf}
\alias{refseq}
\title{refseq, alliance, pubmed, manual, bugsigdb}
\usage{
alliance(
  geneList,
  alliancePath = "GENE-DESCRIPTION-TSV_HUMAN.tsv",
  keyType = "SYMBOL",
  excludeFreq = 2000,
  exclude = "frequency",
  filterMax = FALSE,
  excludeType = ">",
  tfidf = FALSE,
  genePlotNum = 10,
  preserve = FALSE,
  takeMax = FALSE,
  additionalRemove = NA,
  onlyCorpus = FALSE,
  madeUpper = c("dna", "rna"),
  organism = 9606,
  pal = c("blue", "red"),
  numWords = 30,
  scaleRange = c(5, 10),
  autoScale = FALSE,
  showLegend = FALSE,
  orgDb = org.Hs.eg.db,
  edgeLabel = FALSE,
  naEdgeColor = "grey50",
  cooccurrence = FALSE,
  pvclAlpha = 0.95,
  cl = FALSE,
  ngram = 1,
  plotType = "network",
  onlyTDM = FALSE,
  stem = FALSE,
  colorText = FALSE,
  corThresh = 0.2,
  genePlot = FALSE,
  autoThresh = TRUE,
  genePathPlot = NULL,
  genePathPlotSig = 0.05,
  tag = "none",
  layout = "nicely",
  edgeLink = TRUE,
  deleteZeroDeg = TRUE,
  enrich = NULL,
  topPath = 10,
  tagWhole = FALSE,
  mergeCorpus = NULL,
  numOnly = TRUE,
  madeUpperGenes = TRUE,
  onWholeDTM = FALSE,
  pre = TRUE,
  takeMean = FALSE,
  tagPalette = NULL,
  collapse = FALSE,
  addFreqToGene = FALSE,
  useUdpipe = FALSE,
  normalize = FALSE,
  fontFamily = "sans",
  udpipeModel = "english-ewt-ud-2.5-191206.udpipe",
  scaleFreq = NULL,
  colorize = FALSE,
  geneColor = "grey",
  argList = list(),
  useggwordcloud = TRUE,
  wcScale = 10,
  catColors = NULL,
  discreteColorWord = FALSE,
  useSeed = 42,
  scaleEdgeWidth = c(1, 3),
  filterByGO = FALSE,
  docsum = FALSE,
  absolute = TRUE,
  corOption = list()
)

bugsigdb(
  mbList,
  excludeFreq = 1000,
  exclude = "frequency",
  excludeType = ">",
  normalize = FALSE,
  takeMean = FALSE,
  additionalRemove = NA,
  tfidf = FALSE,
  target = "title",
  apiKey = NULL,
  takeMax = FALSE,
  pre = FALSE,
  pvclAlpha = 0.95,
  numOnly = TRUE,
  madeUpper = c("dna", "rna"),
  redo = NULL,
  fontFamily = "sans",
  pal = c("blue", "red"),
  numWords = 15,
  preserve = FALSE,
  metab = NULL,
  metThresh = 0.2,
  curate = TRUE,
  abstArg = list(),
  tagPalette = NULL,
  metCol = NULL,
  scaleRange = c(5, 10),
  showLegend = FALSE,
  ecPlot = FALSE,
  edgeLabel = FALSE,
  mbPlot = FALSE,
  onlyTDM = FALSE,
  ecFile = NULL,
  upTaxFile = NULL,
  filterMax = FALSE,
  mbColor = "grey",
  useUdpipe = FALSE,
  colorize = FALSE,
  cooccurrence = FALSE,
  udpipeModel = "english-ewt-ud-2.5-191206.udpipe",
  scaleFreq = NULL,
  ngram = 1,
  plotType = "network",
  disPlot = FALSE,
  onWholeDTM = FALSE,
  naEdgeColor = "grey50",
  useggwordcloud = TRUE,
  wcScale = 10,
  addFreqToMB = FALSE,
  catColors = NULL,
  useSeed = 42,
  discreteColorWord = FALSE,
  colorText = FALSE,
  corThresh = 0.2,
  tag = "none",
  tagWhole = FALSE,
  stem = FALSE,
  layout = "nicely",
  edgeLink = TRUE,
  deleteZeroDeg = TRUE,
  cl = FALSE,
  autoThresh = TRUE,
  argList = list(),
  docsum = FALSE,
  absolute = TRUE,
  corOption = list(),
  cache = TRUE
)

manual(
  df,
  madeUpper = NULL,
  useFil = NA,
  filType = "above",
  cooccurrence = FALSE,
  filNum = 0,
  useQuanteda = FALSE,
  quantedaArgs = list(),
  pvclAlpha = 0.95,
  numOnly = TRUE,
  tfidf = FALSE,
  cl = FALSE,
  pal = c("blue", "red"),
  numWords = 30,
  scaleRange = c(5, 10),
  scaleFreq = NULL,
  showLegend = FALSE,
  plotType = "network",
  colorText = FALSE,
  corThresh = 0.2,
  layout = "nicely",
  tag = "none",
  tagWhole = FALSE,
  onlyCorpus = FALSE,
  onlyTDM = FALSE,
  queryColor = "grey",
  edgeLabel = FALSE,
  edgeLink = TRUE,
  ngram = 1,
  colorize = FALSE,
  tagPalette = NULL,
  preserve = FALSE,
  takeMax = FALSE,
  catColors = NULL,
  deleteZeroDeg = TRUE,
  additionalRemove = NA,
  naEdgeColor = "grey50",
  normalize = FALSE,
  takeMean = FALSE,
  queryPlot = FALSE,
  collapse = FALSE,
  onWholeDTM = FALSE,
  stem = FALSE,
  argList = list(),
  useUdpipe = FALSE,
  discreteColorWord = FALSE,
  autoThresh = TRUE,
  useggwordcloud = TRUE,
  wcScale = 10,
  fontFamily = "sans",
  addFreqToNonWords = FALSE,
  filterByGO = FALSE,
  docsum = FALSE,
  absolute = TRUE,
  corOption = list(),
  udpipeModel = "english-ewt-ud-2.5-191206.udpipe",
  useSeed = 42
)

pubmed(
  queries,
  useRawQuery = FALSE,
  redo = NULL,
  madeUpper = c("dna", "rna"),
  target = "abstract",
  useFil = NA,
  filType = "above",
  filNum = 0,
  sortOrder = "relevance",
  fontFamily = "sans",
  pvclAlpha = 0.95,
  numOnly = TRUE,
  delim = "OR",
  apiKey = NULL,
  tfidf = FALSE,
  cl = FALSE,
  autoThresh = TRUE,
  pal = c("blue", "red"),
  numWords = 30,
  scaleRange = c(5, 10),
  showLegend = FALSE,
  plotType = "network",
  colorText = FALSE,
  quote = FALSE,
  corThresh = 0.2,
  layout = "nicely",
  tag = "none",
  tagWhole = FALSE,
  onlyCorpus = FALSE,
  onlyTDM = FALSE,
  retMax = 10,
  edgeLabel = FALSE,
  edgeLink = TRUE,
  ngram = 1,
  genePlot = FALSE,
  scaleFreq = NULL,
  onlyDf = FALSE,
  tagPalette = NULL,
  preserve = FALSE,
  takeMax = FALSE,
  catColors = NULL,
  perQuery = FALSE,
  discreteColorWord = FALSE,
  useUdpipe = FALSE,
  udpipeOnlyFreq = FALSE,
  udpipeOnlyFreqNB = FALSE,
  addFreqToQuery = FALSE,
  naEdgeColor = "grey50",
  cooccurrence = FALSE,
  colorize = FALSE,
  queryColor = "grey",
  useggwordcloud = TRUE,
  wcScale = 10,
  distinguish_query = TRUE,
  useSeed = 42,
  udpipeModel = "english-ewt-ud-2.5-191206.udpipe",
  normalize = FALSE,
  takeMean = FALSE,
  absolute = TRUE,
  corOption = list(),
  deleteZeroDeg = TRUE,
  additionalRemove = NA,
  orgDb = org.Hs.eg.db,
  onlyGene = FALSE,
  filterByGO = FALSE,
  docsum = FALSE,
  pre = FALSE,
  onWholeDTM = FALSE,
  madeUpperGenes = TRUE,
  stem = FALSE,
  argList = list(),
  dateRange = NULL,
  cc0 = FALSE
)

refseq(
  geneList,
  keyType = "SYMBOL",
  excludeFreq = 2000,
  exclude = "frequency",
  filterMax = FALSE,
  excludeType = ">",
  tfidf = FALSE,
  genePlotNum = 10,
  preserve = FALSE,
  takeMax = FALSE,
  additionalRemove = NA,
  onlyCorpus = FALSE,
  madeUpper = c("dna", "rna"),
  organism = 9606,
  pal = c("blue", "red"),
  numWords = 30,
  scaleRange = c(5, 10),
  autoScale = FALSE,
  showLegend = FALSE,
  orgDb = org.Hs.eg.db,
  edgeLabel = FALSE,
  naEdgeColor = "grey50",
  cooccurrence = FALSE,
  pvclAlpha = 0.95,
  cl = FALSE,
  multiVals = "first",
  ngram = 1,
  plotType = "network",
  onlyTDM = FALSE,
  stem = FALSE,
  colorText = FALSE,
  corThresh = 0.2,
  genePlot = FALSE,
  autoThresh = TRUE,
  autoNumWords = FALSE,
  genePathPlot = NULL,
  genePathPlotSig = 0.05,
  tag = "none",
  layout = "nicely",
  edgeLink = TRUE,
  deleteZeroDeg = TRUE,
  enrich = NULL,
  topPath = 10,
  ora = FALSE,
  tagWhole = FALSE,
  mergeCorpus = NULL,
  numOnly = TRUE,
  madeUpperGenes = TRUE,
  onWholeDTM = FALSE,
  pre = TRUE,
  takeMean = FALSE,
  tagPalette = NULL,
  collapse = FALSE,
  addFreqToGene = FALSE,
  useUdpipe = FALSE,
  normalize = FALSE,
  fontFamily = "sans",
  udpipeModel = "english-ewt-ud-2.5-191206.udpipe",
  scaleFreq = NULL,
  colorize = FALSE,
  geneColor = "grey",
  argList = list(),
  useggwordcloud = TRUE,
  wcScale = 10,
  catColors = NULL,
  discreteColorWord = FALSE,
  useSeed = 42,
  scaleEdgeWidth = c(1, 3),
  splitByEA = NULL,
  filterByGO = FALSE,
  docsum = FALSE,
  absolute = TRUE,
  corOption = list()
)
}
\arguments{
\item{geneList}{gene ID list}

\item{alliancePath}{path to The Alliance of Genome Resources gene description file.
default to "GENE-DESCRIPTION-TSV_HUMAN.tsv"}

\item{keyType}{default to SYMBOL}

\item{excludeFreq}{default to 5000}

\item{exclude}{"frequency" or "tfidf",}

\item{filterMax}{Use pre-calculated filter based on max-values when excluding TfIdf
Otherwise take sum.}

\item{excludeType}{">" or "<", combined with `exclude` and `excludeFreq`,
e.g. filter the words with the pre-calculated frequency > 5000}

\item{tfidf}{use TfIdf when making TDM, default to FALSE.}

\item{genePlotNum}{number of genes to be plotted (default: 10)}

\item{preserve}{Try to preserve original characters.}

\item{takeMax}{Take max values for each term in term-document matrix}

\item{additionalRemove}{specific words to be excluded}

\item{onlyCorpus}{return only corpus (tm).}

\item{madeUpper}{make these words uppercase in resulting plot,
default to c("rna" and "dna")}

\item{organism}{organism ID to use in `GeneSummary`}

\item{pal}{palette for color gradient in correlation network.
should be a vector of length two like c("red","blue").}

\item{numWords}{the number of words to be shown in the plot.
When `autoThresh` is TRUE, the number of this value will be shown.}

\item{scaleRange}{scale for label and node size in the network.}

\item{autoScale}{scale the label and node size automatically for the large network.}

\item{showLegend}{whether to show legend in the network}

\item{orgDb}{the database used to convert identifiers, default to org.Hs.eg.db.}

\item{edgeLabel}{if TRUE, plot the edge label (default: FALSE)}

\item{naEdgeColor}{edge colors for NA values (linking query with the category other than text)}

\item{cooccurrence}{default to FALSE, if TRUE, use cooccurrence instead of correlation.}

\item{pvclAlpha}{alpha value for pvpick()}

\item{cl}{for parPvclust, parallel clustering can be performed}

\item{ngram}{N-gram specification, default to 1.}

\item{plotType}{"wc" or "network", default to "network"}

\item{onlyTDM}{return only TDM (tm).}

\item{stem}{whether to use stemming when making corpus.}

\item{colorText}{color text label based on frequency in the network}

\item{corThresh}{the correlation (cooccurrence) threshold.}

\item{genePlot}{plot associated genes (default: FALSE)
This option first calculates the high-frequent words, and subsequently calculates
the occurrences of these words in each gene, and prioritize the genes by them.}

\item{autoThresh}{automatically choose thresholding value to show the `numWords`,
when deleteZeroDeg (deleting no-connected words) is TRUE, which is default.}

\item{genePathPlot}{plot associated genes and pathways (default: NULL)
Must be "kegg" or "reactome", automatically set genePlot to TRUE.}

\item{genePathPlotSig}{threshold for adjusted p-values (default: 0.05)}

\item{tag}{perform pvclust on words and colorlize them in wordcloud or network
argument of those accepted in pvclust `method.dist` option, like "correlation".
Default to "none", which performs no tagging.}

\item{layout}{the layout for the network, defaul to "nicely".
It can be one of the layouts implemented in `igraph` and `ggraph`, such as
`kk` (Kamada-Kawai), `nicely` (automatic selection of algorithm), `drl` (the force-directed DrL layout).
The options are available at: https://igraph.org/r/doc/layout_.html}

\item{edgeLink}{if FALSE, use geom_edge_diagonal. if TRUE, geom_edge_link. Default to TRUE.}

\item{deleteZeroDeg}{delete zero degree node from plot in the network}

\item{enrich}{currently, only 'reactome' and 'kegg' is supported.}

\item{topPath}{how many pathway descriptions are included in text analysis,
sorted by p-values in the results.}

\item{tagWhole}{whether to perform pvclust on whole matrix or subset of the matrix.}

\item{mergeCorpus}{specify multiple corpus if intend to combine them.
like PubMed information and RefSeq summary}

\item{numOnly}{delete number only (not deleting XXX123, but delete only the number)}

\item{madeUpperGenes}{make genes upper case automatically (default to TRUE)
This uses the `SYMBOL` key in `orgDb`.}

\item{onWholeDTM}{calculate correlation network
on whole dataset or top-words specified by numWords.}

\item{pre}{remove preset filtering words.}

\item{takeMean}{take mean values for each term in term-document matrix.}

\item{tagPalette}{tag palette when `tag` is TRUE. It is also used for dependency network
using udpipe, and tagging colorization for word cloud.
Default to NULL, which indicates automatically set.}

\item{collapse}{default to FALSE, collapse all the sentences.}

\item{addFreqToGene}{add pseudo frequency to gene in genePlot, default to FALSE.}

\item{useUdpipe}{use udpipe to make a dependency network.}

\item{normalize}{sum normalize the term frequency document-wise.}

\item{fontFamily}{font family to use, default to "sans".}

\item{udpipeModel}{udpipe model file name.}

\item{scaleFreq}{default to NULL, scale the value if specified}

\item{colorize}{color the word nodes by their frequency, and the other nodes by their category.
if colorize=FALSE and addFreqToGene=TRUE, gene nodes are colorized according to the minimum frequency 
of the words in the network}

\item{geneColor}{color for associated genes with words (used when tag or colorize option is TRUE)}

\item{argList}{parameters to pass to wordcloud() or ggwordcloud()}

\item{useggwordcloud}{default to TRUE, otherwise use `wordcloud` function.}

\item{wcScale}{scaling size for ggwordcloud}

\item{catColors}{colors for words and texts when colorize is TRUE and discreteColorWord is TRUE}

\item{discreteColorWord}{colorize words by "Words" category, not frequency.}

\item{useSeed}{random seed}

\item{scaleEdgeWidth}{scale for edge width}

\item{filterByGO}{filter the results to the words obtained from GO terms,
while preserving the number of words to be shown}

\item{docsum}{if TRUE, convert the term-document matrix to binary.}

\item{absolute}{calculate absolute correlation value}

\item{corOption}{passed to `cor` function, like list("method"="kendall")}

\item{mbList}{microbe ID list}

\item{target}{"title" or "abstract"}

\item{apiKey}{api key for eutilities}

\item{redo}{if plot in other parameters, input the previous list}

\item{metab}{tibble of metabolite - taxon association}

\item{metThresh}{threshold of association}

\item{curate}{include articles in bugsigdb}

\item{abstArg}{passed to PubMed function when using curate=FALSE}

\item{metCol}{metabolite data frame column name in the order of
"candidate taxon", "metabolite", "quantitative values for thresholding"}

\item{ecPlot}{plot link between enzyme and microbes
this option requires two files to be passed to enzyme() and getUPTax().}

\item{mbPlot}{plot microbe names}

\item{ecFile}{enzyme database file}

\item{upTaxFile}{UniProt taxonomy file}

\item{mbColor}{color for Microbes when tagPalette or catColors 
is not specified}

\item{disPlot}{plot diseases}

\item{addFreqToMB}{add pseudo frequency to microbes in mbPlot}

\item{cache}{Caching when BugSigDB is downloaded (default to TRUE)}

\item{df}{manual document data.frame (must have column `text`) or vector of text.
If `query` column and other columns are present, regards them as category related to the text on the same row.}

\item{useFil}{filter based on "GS_TfIdf" (whole gene summary tf-idf)
or "BSDB_TfIdf" (whole bugsigdb tf-idf)}

\item{filType}{"above" or "below"}

\item{filNum}{specify filter tfidf}

\item{useQuanteda}{use quanteda functions to generate}

\item{quantedaArgs}{list of arguments to be passed to tokens()}

\item{queryColor}{color for associated queries with words}

\item{queryPlot}{plot the query in the graph in relation with the words}

\item{addFreqToNonWords}{add pseudo-frequency corresponding to minimum
frequency of the words to nodes other than words}

\item{queries}{query ID list}

\item{useRawQuery}{if you would like to send the query as is, please set this option to TRUE.}

\item{sortOrder}{sort order, passed to rentrez function}

\item{delim}{delimiter for queries}

\item{quote}{whether to quote the queries}

\item{retMax}{how many items are to be retlieved?}

\item{onlyDf}{return only the raw data.frame of searching PubMed}

\item{perQuery}{search for the queries one by one recursively, not using `delim`.}

\item{udpipeOnlyFreq}{when using udpipe, include only high-frequent words}

\item{udpipeOnlyFreqNB}{when using udpipe, include only the neighbors of
high-frequent words}

\item{addFreqToQuery}{add pseudo-frequency to query node}

\item{distinguish_query}{if TRUE, distinguish query and returned texts
by appending (Q) on query}

\item{onlyGene}{plot only the gene symbol
(orgDb with SYMBOL key can be used)}

\item{dateRange}{if specified, restrict the range of publication date.
Must be the two-length vector, like `c("2013/1/1", "2023/1/1")`}

\item{cc0}{effective only in `pubmed` function. The prefetched data in `pubmedMini`
package is used for gene query.}

\item{multiVals}{passed to mapIds function when converting ENTREZID to SYMBOL when
genePlot is on.}

\item{autoNumWords}{determine the number of words to be shown by ORA, default to FALSE.}

\item{ora}{perform over-representation analysis or not (experimental)}

\item{splitByEA}{automatically split the genes based on significant enrichment analysis results,
and returns the list of object for each term. Default to NULL. Must be 'kegg' or 'reactome',
in which the function performs over-representation analysis by enrichKEGG or enrichPathway in
clusterProfiler and ReactomePA.}
}
\value{
`biotext` class object
}
\description{
Text mining RefSeq description, PubMed, BugSigDB
and the other manually curated textual data.
}
\details{
The main functions of the \pkg{biotextgraph} package. The functions accepts
a character vector of biological entities (such as gene identifiers)
and returns the summarized statistics and visualization
contained in \code{biotext} object.
}
\examples{
geneList <- c("DDX41","PNKP","ERCC1","IRF3","XRCC1")
\dontrun{alliance(geneList)}
mbList <- c("Veillonella dispar")
\dontrun{
    bugsigdb(mbList, plotType="wc")
}
ret <- refseq("DDX41", plotType="wc")
manual(getSlot(ret, "rawText")$Gene_summary, plotType="wc")
\dontrun{pubmed("DDX41")}
geneList <- c("DDX41","PNKP","ERCC1","IRF3","XRCC1")
refseq(geneList)

}
