% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_sc_wordcloud.R
\name{geom_sc_wordcloud}
\alias{geom_sc_wordcloud}
\title{geom_sc_wordcloud}
\usage{
geom_sc_wordcloud(
  markers,
  show_markers = NULL,
  gene_name = FALSE,
  p_val_adj_threshold = 0.05,
  keyType = "SYMBOL",
  rad = NULL,
  wc_alpha = 1,
  withTitle = FALSE,
  rot.per = 0.4,
  random.order = FALSE,
  use_shadowtext = TRUE,
  bg.colour = "white",
  wcScale = 4,
  args = list(),
  wcArgs = list(),
  sortBy = "avg_log2FC",
  scaleNumber = 2,
  decreasing = TRUE,
  geneNum = 50,
  base_ellipse = FALSE,
  base_dens = FALSE,
  withggfx = NULL,
  ggfxParams = list()
)
}
\arguments{
\item{markers}{FindAllMarkers() results}

\item{show_markers}{candidate clusters to be appear in plot, default to NULL,
which means all the clusters are plotted}

\item{gene_name}{show gene names instead of textual information}

\item{p_val_adj_threshold}{default to 0.05, used in subsetting marker data frame
using \code{p_val_adj}}

\item{keyType}{key type of those listed in the markers. default to SYMBOL.}

\item{rad}{named vector of radius for wordclouds. if specified, this value is 
used to define positions of annotation_custom.}

\item{wc_alpha}{alpha value for word cloud}

\item{withTitle}{whether to show title in each grob of wordcloud}

\item{rot.per}{ggwordcloud parameter}

\item{random.order}{ggwordcloud parameter}

\item{use_shadowtext}{use shadowtext for wordcloud}

\item{bg.colour}{the background color of shadowtext}

\item{wcScale}{word cloud scaling factor}

\item{args}{passed to wcGeneSummary()}

\item{wcArgs}{passed to ggwordcloud()}

\item{sortBy}{default to avg_log2FC, "log10p" can be specified.}

\item{scaleNumber}{scale the frequency of words by this number
in `gene_name`}

\item{decreasing}{sort by decreasing order or not}

\item{geneNum}{number of genes to be included in wordclouds}

\item{base_ellipse}{if TRUE, wordclouds are placed based on \code{stat_ellipse}.}

\item{base_dens}{if TRUE, wordclouds are placed based on density}

\item{withggfx}{applying ggfx filters}

\item{ggfxParams}{parameter list for ggfx}
}
\value{
ggplot
}
\description{
geom_sc_wordcloud
}
