% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotextgraph.R
\name{getSlot}
\alias{getSlot}
\title{getSlot}
\usage{
getSlot(x, slot)
}
\arguments{
\item{x}{biotext object}

\item{slot}{slot name}
}
\value{
attribute value
}
\description{
get the slot value from biotext object
}
