% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getUPtax}
\alias{getUPtax}
\title{getUPtax}
\usage{
getUPtax(file, candUP, candTax = NULL)
}
\arguments{
\item{file}{downloaded file}

\item{candUP}{candidate UniProt identifiers}

\item{candTax}{candidate taxonomy name}
}
\description{
Obtain the list of UniProt taxonomy ID.
https://ftp.uniprot.org/pub/databases/uniprot/current_release/knowledgebase/complete/docs/speclist.txt
The file above must be specified to file argument.
}
