% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWordsOnDendro.R
\name{getWordsOnDendro}
\alias{getWordsOnDendro}
\title{getWordsOnDendro}
\usage{
getWordsOnDendro(
  dhc,
  geneVec,
  geneNumLimit = 1000,
  geneVecType = "ENSEMBL",
  excludeTfIdf = NA,
  numberOfWords = 25,
  excludeFreq = 5000,
  orgDb = org.Hs.eg.db,
  tfidf = TRUE,
  madeUpper = c("rna", "dna")
)
}
\arguments{
\item{dhc}{dendrogram}

\item{geneVec}{gene-named vector of node names in dendrogram}

\item{geneNumLimit}{when the gene number is above this threshold,
pyramid plots are not produced}

\item{geneVecType}{type of the name of geneVec (default: ENSEMBL)}

\item{excludeTfIdf}{filter using tfidf}

\item{numberOfWords}{the number of words to plot (default: 25)}

\item{excludeFreq}{words with the frequency above
this threshold is excluded beforehand}

\item{orgDb}{organism database}

\item{tfidf}{use tfidf when making TDM}

\item{madeUpper}{words with uppercase}
}
\value{
list of pyramid plot grobs and its positions
}
\description{
Get grobs to plot on the dendrogram with the position information
}
\examples{
mod <- returnExample()
result <- hclust(dist(t(mod$MEs)))
dhc <- result |>
    as.dendrogram() |>
    dendextend::hang.dendrogram()
geneVec <- paste0("ME", mod$colors)
names(geneVec) <- names(mod$colors)
getWordsOnDendro(dhc, geneVec, numberOfWords = 2)
}
