% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{graph_cluster}
\alias{graph_cluster}
\title{graph_cluster}
\usage{
graph_cluster(ret, func = igraph::cluster_leiden, factorize = TRUE)
}
\arguments{
\item{ret}{biotext class object}

\item{func}{community detection algorithm in igraph}

\item{factorize}{convert to factor upon assigning}
}
\value{
biotext class object
}
\description{
Performs graph-based clustering.
}
\details{
Performs graph-based clustering using igraph and stores the information
in the biotext object. Default to use `igraph::cluster_leiden`.
}
\examples{
testgenes <- c("IRF3","PNKP","DDX41","ERCC1","ERCC2","XRCC1")
refseq(testgenes) |> graph_cluster()
}
