% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{make_corpus}
\alias{make_corpus}
\title{make_corpus}
\usage{
make_corpus(
  ret,
  collapse = FALSE,
  num_only = TRUE,
  stem = FALSE,
  preserve = TRUE,
  ngram = 1
)
}
\arguments{
\item{ret}{biotext class object}

\item{collapse}{collapse all the sentences to one sentence}

\item{num_only}{delete number only (not deleting XXX123)}

\item{stem}{stem the words or not (default to FALSE)}

\item{preserve}{preserve the original cases}

\item{ngram}{n-gram}
}
\value{
biotext class object
}
\description{
Make corpus based on biotext class
}
\details{
Make corpus based on biotext class using tm.
Ngram, stemming parameters can be specified by this function.
}
\examples{
obtain_refseq("PNKP") |> make_corpus()
}
