% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{make_graph}
\alias{make_graph}
\title{make_graph}
\usage{
make_graph(
  ret,
  num_words = 30,
  cor_threshold = 0.2,
  cooccurrence = FALSE,
  whole = FALSE,
  bn = FALSE,
  R = 20
)
}
\arguments{
\item{ret}{biotext class object}

\item{num_words}{number of words to include}

\item{cor_threshold}{correlation threshold}

\item{cooccurrence}{use cooccurrence}

\item{whole}{return correlation or cooccurrence for whole words}

\item{bn}{perform Bayesian network inference}

\item{R}{parameter for boot.strength function}
}
\value{
biotext class object
}
\description{
make correlation or cooccurrence network
}
