% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{make_TDM}
\alias{make_TDM}
\title{make_TDM}
\usage{
make_TDM(
  ret,
  tfidf = FALSE,
  normalize = FALSE,
  takeMean = FALSE,
  takeMax = FALSE
)
}
\arguments{
\item{ret}{biotext class}

\item{tfidf}{use TF-IDF}

\item{normalize}{normalize the values to document number}

\item{takeMean}{take the mean value for words to rank}

\item{takeMax}{take the max value for words to rank}
}
\value{
biotext class object
}
\description{
make_TDM
}
\examples{
obtain_refseq("PNKP") |> make_corpus() |> make_TDM()
}
