% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeBar}
\alias{makeBar}
\title{makeBar}
\usage{
makeBar(
  queries,
  top = 10,
  keyType = "SYMBOL",
  pal = NULL,
  textSize = 20,
  reorder = TRUE
)
}
\arguments{
\item{queries}{gene IDs}

\item{top}{how many numbers of words to be shown}

\item{keyType}{default to SYMBOL}

\item{pal}{palette used in barplot}

\item{textSize}{text size in barplot}

\item{reorder}{order by frequency or not}
}
\value{
barplot of word frequency
}
\description{
Makeing a barplot of word frequency from queried genes
}
\examples{
geneList <- c("DDX41")
makeBar(geneList)
}
