% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWordsOnDendro.R
\name{makeCorpus}
\alias{makeCorpus}
\title{makeCorpus}
\usage{
makeCorpus(docs, filterWords, additionalRemove, numOnly)
}
\arguments{
\item{docs}{corpus to clean}

\item{filterWords}{words to filter based on frequency}

\item{additionalRemove}{words to filter}

\item{numOnly}{delete number only}
}
\value{
cleaned corpus
}
\description{
Clean-up the corpus
}
