% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual.R
\name{manual}
\alias{manual}
\title{manual}
\usage{
manual(
  df,
  madeUpper = NULL,
  useFil = NA,
  filType = "above",
  cooccurrence = FALSE,
  filNum = 0,
  useQuanteda = FALSE,
  quantedaArgs = list(),
  pvclAlpha = 0.95,
  numOnly = TRUE,
  tfidf = FALSE,
  cl = FALSE,
  pal = c("blue", "red"),
  numWords = 30,
  scaleRange = c(5, 10),
  scaleFreq = NULL,
  showLegend = FALSE,
  plotType = "network",
  colorText = FALSE,
  corThresh = 0.2,
  layout = "nicely",
  tag = "none",
  tagWhole = FALSE,
  onlyCorpus = FALSE,
  onlyTDM = FALSE,
  bn = FALSE,
  R = 20,
  queryColor = "grey",
  edgeLabel = FALSE,
  edgeLink = TRUE,
  ngram = NA,
  colorize = FALSE,
  tagPalette = NULL,
  preserve = TRUE,
  takeMax = FALSE,
  catColors = NULL,
  deleteZeroDeg = TRUE,
  additionalRemove = NA,
  naEdgeColor = "grey50",
  normalize = FALSE,
  takeMean = FALSE,
  queryPlot = FALSE,
  collapse = FALSE,
  onWholeDTM = FALSE,
  stem = FALSE,
  argList = list(),
  useUdpipe = FALSE,
  discreteColorWord = FALSE,
  useggwordcloud = TRUE,
  wcScale = 10,
  fontFamily = "sans",
  addFreqToNonWords = FALSE,
  udpipeModel = "english-ewt-ud-2.5-191206.udpipe",
  useSeed = 42
)
}
\arguments{
\item{df}{df}

\item{madeUpper}{make the words uppercase in resulting plot}

\item{useFil}{filter based on "GS_TfIdf" (whole gene summary tf-idf)
or "BSDB_TfIdf" (whole bugsigdb tf-idf)}

\item{filType}{"above" or "below"}

\item{cooccurrence}{default to FALSE, if TRUE, use cooccurrence instead of correlation}

\item{filNum}{specify filter tfidf}

\item{useQuanteda}{use quanteda functions to generate}

\item{quantedaArgs}{list of arguments to be passed to tokens()}

\item{pvclAlpha}{alpha for pvpick()}

\item{numOnly}{delete number only}

\item{tfidf}{use TfIdf when making TDM}

\item{cl}{cluster to pass to pvclust (snow::makeCluster(n))}

\item{pal}{palette for color gradient in correlation network}

\item{numWords}{the number of words to be shown}

\item{scaleRange}{scale for label and node size in correlation network}

\item{scaleFreq}{scale the frequency}

\item{showLegend}{whether to show legend in correlation network}

\item{plotType}{"wc" or "network"}

\item{colorText}{color text label based on frequency in correlation network}

\item{corThresh}{the correlation threshold}

\item{layout}{the layout for correlation network, defaul to "nicely"}

\item{tag}{perform pvclust on words and colorlize them in wordcloud or network
argument of "cor" or "tdm". Default to "none", which performs no tagging.
If wordcloud, tagging will be performed on TDM.}

\item{tagWhole}{tag based on whole data or subset}

\item{onlyCorpus}{return only corpus}

\item{onlyTDM}{return only TDM, if quanteda, return DFM}

\item{bn}{perform bootstrap-based Bayesian network inference 
instead of correlation using bnlearn}

\item{R}{how many bootstrap when bn is stated}

\item{queryColor}{color for query in the network plot
if tagPalette or catColors is NULL, determined palettes with 
queryColor replaced with this color will be used.}

\item{edgeLabel}{if TRUE, plot the edge label (default: FALSE)}

\item{edgeLink}{if FALSE, use geom_edge_diagonal}

\item{ngram}{default to NA (1)}

\item{colorize}{color the nodes and texts based on their category}

\item{tagPalette}{palette for each tag when tag is TRUE, default to NULL
(e.g. `cluster_1`, `cluster_2`, ...)}

\item{preserve}{try to preserve the original characters}

\item{takeMax}{when summarizing term-document matrix, take max.
Otherwise take sum.}

\item{catColors}{named vector showing colors for each category
(e.g. `query`, `Words` and specified column names in data.frame)}

\item{deleteZeroDeg}{delete zero degree node from plot in correlation network}

\item{additionalRemove}{specific words to be excluded}

\item{naEdgeColor}{edge color values for NA}

\item{normalize}{sum normalize the term frequency document-wise}

\item{takeMean}{take mean values for each term in term-document matrix}

\item{queryPlot}{plot query}

\item{collapse}{default to FALSE, collapse all the sentences}

\item{onWholeDTM}{calculate correlation network
on whole dataset or top-words specified by numWords}

\item{stem}{whether to use stemming}

\item{argList}{parameters to pass to wordcloud()}

\item{useUdpipe}{use udpipe to make a network}

\item{discreteColorWord}{colorize words by "Words" category, not frequency.}

\item{useggwordcloud}{default to TRUE}

\item{wcScale}{scaling size for ggwordcloud}

\item{fontFamily}{font family to use, default to "sans"}

\item{addFreqToNonWords}{add pseudo-frequency corresponding to minimum
frequency of the words to nodes other than words}

\item{udpipeModel}{udpipe model file name}

\item{useSeed}{seed}
}
\value{
list of data frame and ggplot2 object
}
\description{
Produce networks using manual input
}
\examples{
ret <- refseq("DDX41", plotType="wc")
manual(ret@rawText$Gene_summary, plotType="wc")
}
