% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{obtain_alliance}
\alias{obtain_alliance}
\title{obtain_alliance}
\usage{
obtain_alliance(
  geneList,
  file = "GENE-DESCRIPTION-TSV_HUMAN.tsv",
  keyType = "SYMBOL",
  org_db = org.Hs.eg.db
)
}
\arguments{
\item{geneList}{gene ID list}

\item{file}{filename, default to "GENE-DESCRIPTION-TSV_HUMAN.tsv"}

\item{keyType}{key type of gene list}

\item{org_db}{used to convert ID}
}
\value{
biotext class object
}
\description{
obtain gene description from 
The Alliance of Genome Resources
https://www.alliancegenome.org/downloads
}
\examples{
\dontrun{obtain_alliance(c("PNKP"))}
}
