% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{obtain_bugsigdb}
\alias{obtain_bugsigdb}
\title{obtain_bugsigdb}
\usage{
obtain_bugsigdb(mb_list, target = "title", curate = TRUE, api_key = NULL)
}
\arguments{
\item{mb_list}{microbe list}

\item{target}{title or abstract}

\item{curate}{curated articles (TRUE) or search for PubMed (FALSE)
if FALSE, same for fetching from PubMed by specified query}

\item{api_key}{API key for PubMed}
}
\value{
biotext class object
}
\description{
obtain microbe-related information from bugsigdb
}
