% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{obtain_manual}
\alias{obtain_manual}
\title{obtain_manual}
\usage{
obtain_manual(df)
}
\arguments{
\item{df}{data frame}
}
\description{
obtain biotext class object from manually specified data
}
\examples{
obtain_manual(data.frame(text=c("test")))

}
