% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{obtain_pubmed}
\alias{obtain_pubmed}
\title{obtain_pubmed}
\usage{
obtain_pubmed(
  queries,
  target = "title",
  delim = "OR",
  api_key = NULL,
  ret_max = 10,
  limit = 10,
  sort_order = "relevance",
  quote = FALSE
)
}
\arguments{
\item{queries}{query list}

\item{target}{title or abstract}

\item{delim}{delimiter for query, default to OR}

\item{api_key}{API key for PubMed}

\item{ret_max}{max articles to get}

\item{limit}{query number limit}

\item{sort_order}{sort by relevance or pubdate}

\item{quote}{whether to quote each query}
}
\value{
biotext class object
}
\description{
Obtain PubMed text data
}
\details{
Obtain pubmed textual data and makes a biotext object.
}
\seealso{
generalf
}
