% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{obtain_refseq}
\alias{obtain_refseq}
\title{obtain_refseq}
\usage{
obtain_refseq(
  geneList,
  keyType = "SYMBOL",
  organism = 9606,
  org_db = org.Hs.eg.db
)
}
\arguments{
\item{geneList}{gene list}

\item{keyType}{key type of gene list}

\item{organism}{organism}

\item{org_db}{used to convert ID}
}
\value{
biotext class object
}
\description{
obtain RefSeq text data
}
