% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_sc_wordcloud.R
\name{obtainMarkersWC}
\alias{obtainMarkersWC}
\title{obtainMarkersWC}
\usage{
obtainMarkersWC(
  markers,
  cols,
  wcArgs,
  eps = 1e-10,
  sort_by = "avg_log2FC",
  decreasing = TRUE,
  scale_number = 10,
  wcScale = 5,
  geneNum = 50
)
}
\arguments{
\item{markers}{marker data frame}

\item{cols}{list of colors}

\item{wcArgs}{arguments for ggwordcloud}

\item{eps}{when taking log of p-values, this value will be added}

\item{sort_by}{default to avg_log2FC, "log10p" can be specified.}

\item{decreasing}{sort by decreasing order or not}

\item{scale_number}{scale the frequency by this number}

\item{wcScale}{scale for scale_size_area()}

\item{geneNum}{number of genes to be included in wordclouds}
}
\description{
obtainMarkersWC
}
