% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_sc_wordcloud.R
\name{obtainMarkersWCScran}
\alias{obtainMarkersWCScran}
\title{obtainMarkersWCScran}
\usage{
obtainMarkersWCScran(
  markers,
  cols,
  wcArgs,
  eps = 1e-10,
  sortBy = "summary.logFC",
  decreasing = TRUE,
  scaleNumber = 10,
  wcScale = 5,
  geneNum = 50,
  withggfx = NULL,
  ggfxParams = list()
)
}
\arguments{
\item{markers}{marker list}

\item{cols}{list of colors}

\item{wcArgs}{arguments for ggwordcloud}

\item{eps}{when taking log of p-values, this value will be added}

\item{sortBy}{default to summary.logFC, "log10p" can be specified.}

\item{decreasing}{sort by decreasing order or not}

\item{scaleNumber}{scale the frequency by this number}

\item{wcScale}{scale for scale_size_area()}

\item{geneNum}{number of genes to be included in wordclouds}

\item{withggfx}{applying ggfx filters}

\item{ggfxParams}{parameter list for ggfx}
}
\value{
list of ggplot containing wordclouds
}
\description{
Make gene wordcloud from scran::findMarkers() results
}
\details{
using the results of marker gene identification such as `findMarkers` from scran,
Recursively summarize the textual information of markers and output the wordclouds.
}
\examples{
df <- data.frame(
  p.value=c(0.01, 0.01),gene=c("PNKP","DDX41")
)
row.names(df) <- df$gene
markers <- list("1"=df)
colors <- list("1"="blue")
obtainMarkersWCScran(markers,
                     cols=colors,
                     wcArgs=list(),
                     sortBy="p.value",
                     geneNum=2)
}
