% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtainTextPosition}
\alias{obtainTextPosition}
\title{obtainTextPosition}
\usage{
obtainTextPosition(ig, sort.by = node_rank_fabric(), verbose = FALSE)
}
\arguments{
\item{ig}{igraph}

\item{sort.by}{the argument to be passed to fabric layout function}

\item{verbose}{show logs}
}
\value{
tbl_graph
}
\description{
obtain text position in biofabric layout
}
\examples{
refseq(c("PNKP","DDX41"))@igraph |> obtainTextPosition()
}
