% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcGeneSummary.R
\name{osp}
\alias{osp}
\title{osp}
\usage{
osp(
  geneList,
  keyType = "SYMBOL",
  excludeFreq = 2000,
  excludeTfIdf = NULL,
  tfidf = FALSE,
  genePlotNum = 10,
  preserve = TRUE,
  additionalRemove = NA,
  onlyCorpus = FALSE,
  madeUpper = c("dna", "rna"),
  organism = 9606,
  pal = c("blue", "red"),
  numWords = 15,
  scaleRange = c(5, 10),
  showLegend = FALSE,
  orgDb = org.Hs.eg.db,
  edgeLabel = FALSE,
  pvclAlpha = 0.95,
  bn = FALSE,
  R = 20,
  cl = FALSE,
  ngram = NA,
  plotType = "wc",
  onlyTDM = FALSE,
  stem = FALSE,
  colorText = FALSE,
  corThresh = 0.2,
  genePlot = FALSE,
  genePathPlot = NA,
  genePathPlotSig = 0.05,
  tag = FALSE,
  layout = "nicely",
  edgeLink = TRUE,
  deleteZeroDeg = TRUE,
  enrich = NULL,
  topPath = 10,
  ora = FALSE,
  tagWhole = FALSE,
  mergeCorpus = NULL,
  numOnly = TRUE,
  madeUpperGenes = TRUE,
  onWholeDTM = FALSE,
  autoFilter = FALSE,
  verbPOS = c("VBZ"),
  nodePal = palette(),
  verb = FALSE,
  udpipeModel = "english-ewt-ud-2.5-191206.udpipe",
  ...
)
}
\description{
alias for wcGeneSummary
}
\examples{
osp("DDX41")
}
