% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcBSDB.R
\name{ospb}
\alias{ospb}
\title{ospb}
\usage{
ospb(
  mbList,
  excludeFreq = 1000,
  excludeTfIdf = NA,
  additionalRemove = NA,
  tfidf = FALSE,
  target = "title",
  apiKey = NULL,
  pre = FALSE,
  pvclAlpha = 0.95,
  numOnly = TRUE,
  madeUpper = c("dna", "rna"),
  redo = NULL,
  pal = c("blue", "red"),
  numWords = 15,
  scaleRange = c(5, 10),
  showLegend = FALSE,
  edgeLabel = FALSE,
  mbPlot = FALSE,
  onlyTDM = FALSE,
  ngram = NA,
  plotType = "wc",
  disPlot = FALSE,
  onWholeDTM = FALSE,
  colorText = FALSE,
  corThresh = 0.2,
  tag = FALSE,
  tagWhole = FALSE,
  layout = "nicely",
  edgeLink = TRUE,
  deleteZeroDeg = TRUE,
  cl = FALSE,
  ...
)
}
\description{
alias for wcBSDB
}
\examples{
ospb("Veillonella dispar")
}
