% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osplot.R
\name{osplot}
\alias{osplot}
\title{osplot}
\usage{
osplot(target, ...)
}
\arguments{
\item{target}{"abstract", "bugsigdb", "refseq"}

\item{...}{passed to each function}
}
\value{
list of data frames and ggplot2 object
}
\description{
wrapper for functions wcGeneSummary, wcAbst, and wcBSDB
}
\examples{
geneList <- c("DDX41")
osplot("refseq", geneList)
}
