% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathviewText.R
\name{pathviewText}
\alias{pathviewText}
\title{pathviewText}
\usage{
pathviewText(
  geneList,
  keyType,
  pid,
  org = "hsa",
  pal = "RdBu",
  target = "refseq",
  searchTerms = NULL,
  node.types = "gene",
  termMap = NULL,
  orgDb = org.Hs.eg.db,
  textSize = 12,
  numWords = 20,
  trans = FALSE,
  areas = NULL,
  argList = list()
)
}
\arguments{
\item{geneList}{gene ID list}

\item{keyType}{keytype for queried gene}

\item{pid}{pathway id}

\item{org}{organism ID in KEGG, default to hsa}

\item{pal}{palette to be passed to RColorBrewer
for coloring of nodes}

\item{target}{target DB for word, "refseq" or "abstract"}

\item{searchTerms}{search terms to be used in wcAbst,
if different than query geneList}

\item{node.types}{passed to node.map e.g. "genes" and "ortholog"}

\item{termMap}{data.frame consisting of columns
"query" and "description", which shows query as "geneList",
and description as "searchTerms".
e.g. "K08097" and "phosphosulfolactate synthase"}

\item{orgDb}{organism database to convert symbol}

\item{textSize}{text size in barplot}

\item{numWords}{number of words in barplot}

\item{trans}{transpose the barplot}

\item{areas}{used in patchwork}

\item{argList}{passed to wc functions}
}
\value{
osplot object, list of plots of pathway, and barplot and concatentaed image
}
\description{
colorize pathview map, and link to word frequency
}
\examples{
query <- c("TP53","CDC45","CDC6")
\dontrun{pathviewText(query, keyType = "SYMBOL", pid = "04110", org = "hsa")}
}
