% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{perform_ora}
\alias{perform_ora}
\title{perform_ora}
\usage{
perform_ora(ret, threshold = 0.05)
}
\arguments{
\item{ret}{biotext class}

\item{threshold}{ORA threshold to filter (bonferroni-corrected p-value)}
}
\value{
biotext class object
}
\description{
perform_ora
}
\examples{
testgenes <- c("IRF3","PNKP","DDX41","ERCC1","ERCC2","XRCC1")
obtain_refseq(testgenes) |> perform_ora()
}
