% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_biofabric}
\alias{plot_biofabric}
\title{plot_biofabric}
\usage{
plot_biofabric(
  res,
  sort.by = node_rank_fabric(),
  size = "rank",
  end_shape = "circle",
  color_map = NULL,
  highlight = NULL,
  highlight_color = "tomato",
  textScaleRange = c(1.5, 2)
)
}
\arguments{
\item{res}{biotext object}

\item{sort.by}{passed to fabric layout}

\item{size}{mapping to use for size of text on x-axis
default to "rank", which uses default ordering,
other accepted options are those in the node table of tbl_graph object}

\item{end_shape}{end shape on `geom_edge_span`}

\item{color_map}{passed to aes mapping in geom_node_range.
Override `highlight` option.}

\item{highlight}{which category to highlight, default to NULL}

\item{highlight_color}{highlight color of the category}

\item{textScaleRange}{decide text scale range on x-axis}
}
\value{
ggplot2
}
\description{
plot the network in biofabric layout
Internally, obtainTextPosition() is used and 
the layers are stacked.
For BioFabric layout, please refer to: https://biofabric.systemsbiology.net/
Citation is: https://doi.org/10.1186/1471-2105-13-275
}
\examples{
refseq(c("DDX41","PNKP")) |> plot_biofabric()


}
