% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{plot_biotextgraph}
\alias{plot_biotextgraph}
\title{plot_biotextgraph}
\usage{
plot_biotextgraph(
  ret,
  edge_link = TRUE,
  edge_label = FALSE,
  show_legend = FALSE,
  cat_colors = NULL,
  query_color = "grey",
  font_family = "sans",
  colorize = FALSE,
  color_by_tag = FALSE,
  color_by_community = FALSE,
  tag_colors = NULL,
  color_text = TRUE,
  scale_range = c(5, 10),
  pal = c("blue", "red"),
  layout = "nicely",
  na_edge_color = "grey",
  add_pseudo_freq = FALSE,
  use_seed = 42
)
}
\arguments{
\item{ret}{biotext object}

\item{edge_link}{whether to use edge link or edge diagonal}

\item{edge_label}{whether to show edge label}

\item{show_legend}{whether to show legend}

\item{cat_colors}{named vector specifying color for each category of node}

\item{query_color}{color for queried nodes}

\item{font_family}{font family}

\item{colorize}{colorize the word frequency and query separately
if FALSE, no node is plotted for query nodes. it overrides color_by_tag.}

\item{color_by_tag}{color by tagging information (color scale will be discrete)}

\item{color_by_community}{color by community information in graph}

\item{tag_colors}{palette for node color in discrete mode}

\item{color_text}{whether to color the text}

\item{scale_range}{scale range for node and text size}

\item{pal}{palette for node color specified as (low, high)}

\item{layout}{graph layout}

\item{na_edge_color}{edge color for NA value}

\item{add_pseudo_freq}{add pseudo value for nodes other than words}

\item{use_seed}{seed value for ggrepel}
}
\value{
biotext class object
}
\description{
plot biotextgraph after processing graph
}
