% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{plot_wordcloud}
\alias{plot_wordcloud}
\title{plot_wordcloud}
\usage{
plot_wordcloud(
  ret,
  num_words = 30,
  pal = palette(),
  make_upper = NULL,
  scale_freq = NULL,
  color_by_tag = FALSE,
  font_family = "sans",
  use_ggwordcloud = TRUE,
  arg_list = list(),
  wc_scale = 10
)
}
\arguments{
\item{ret}{biotext object}

\item{num_words}{number of words to include}

\item{pal}{palette for colorizing words}

\item{make_upper}{make uppercase for these words}

\item{scale_freq}{scale the frequency for visualization}

\item{color_by_tag}{colorize based on tagging information}

\item{font_family}{font family}

\item{use_ggwordcloud}{use ggwordcloud or wordcloud function}

\item{arg_list}{arguments to pass to wordcloud functions}

\item{wc_scale}{scale factor for ggwordcloud}
}
\value{
biotext class object
}
\description{
plot_wordcloud
}
