% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osplot.R
\name{plotORA}
\alias{plotORA}
\title{plotORA}
\usage{
plotORA(x, thresh = 0.001)
}
\arguments{
\item{x}{osplot object}

\item{thresh}{hline to draw in plot}
}
\description{
plot volcano-plot like plot for ORA results
}
