% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.R
\name{plotReducedDimWithTexts}
\alias{plotReducedDimWithTexts}
\title{plotReducedDimWithTexts}
\usage{
plotReducedDimWithTexts(
  sce,
  marker.info,
  colour_by = "label",
  point_alpha = 0.4,
  use_shadowtext = TRUE,
  bg.colour = "white",
  which.label = NULL,
  wc_alpha = 1,
  wcScale = 5,
  rot.per = 0.4,
  r = NULL,
  top = 10,
  random.order = FALSE,
  dimred = "PCA",
  withTitle = FALSE,
  args = list()
)
}
\arguments{
\item{sce}{sce object}

\item{marker.info}{results of findMarkers()}

\item{colour_by}{colorize by this label}

\item{point_alpha}{point alpha}

\item{use_shadowtext}{use shadowtext for wordcloud}

\item{bg.colour}{shadowtext background color}

\item{which.label}{which label to plot text}

\item{wc_alpha}{alpha value for wordcloud}

\item{wcScale}{scaling value for wordcloud}

\item{rot.per}{ggwordcloud parameter}

\item{r}{named vector of size of each cluster}

\item{top}{Top-{top} genes are included}

\item{random.order}{ggwordcloud parameter}

\item{dimred}{dimension reduction method}

\item{withTitle}{whether to append title on wordcloud}

\item{args}{parameters to passed to wcGeneSummary}
}
\value{
single-cell plot with text annotation
}
\description{
plotReducedDimWithTexts
}
