% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotextgraph.R
\name{plotWC}
\alias{plotWC}
\title{plotWC}
\usage{
plotWC(
  x,
  tagPalette = NULL,
  madeUpper = c("dna", "rna"),
  preserve = TRUE,
  scaleFreq = NULL,
  fontFamily = "sans",
  wcScale = 10,
  argList = list(),
  useggwordcloud = TRUE,
  asis = FALSE
)
}
\arguments{
\item{tagPalette}{word-cluster coloring specification}
}
\value{
wordcloud visualization
}
\description{
plot the wordcloud changing the visualization parameters
}
