% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{process_network_gene}
\alias{process_network_gene}
\title{process_network_gene}
\usage{
process_network_gene(
  ret,
  delete_zero_degree = TRUE,
  make_upper = NULL,
  make_upper_gene = TRUE,
  org_db = org.Hs.eg.db,
  gene_plot = FALSE,
  gene_plot_num = 5,
  gene_path_plot = NULL,
  gene_path_plot_threshold = 0.05,
  distinguish_query = FALSE
)
}
\arguments{
\item{ret}{biotext class object}

\item{delete_zero_degree}{exclude zero degree nodes from plotting}

\item{make_upper}{make these words to uppercase}

\item{make_upper_gene}{make gene symbol to uppercase}

\item{org_db}{organism database to convert IDs}

\item{gene_plot}{plot and return associated genes}

\item{gene_plot_num}{how many genes are to be plotted}

\item{gene_path_plot}{plot associated pathways related to genes}

\item{gene_path_plot_threshold}{threshold for pathway enrichment}

\item{distinguish_query}{distinguish query with words in obtained text}
}
\value{
biotext class object
}
\description{
perform tasks before plotting
like obtaining associated genes to plot
}
\examples{
testgenes <- c("IRF3","PNKP","DDX41","ERCC1","ERCC2","XRCC1")
refseq(testgenes) |>
process_network_gene()
}
