% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{process_network_manual}
\alias{process_network_manual}
\title{process_network_manual}
\usage{
process_network_manual(
  ret,
  delete_zero_degree = TRUE,
  make_upper = NULL,
  query_plot = FALSE
)
}
\arguments{
\item{ret}{biotext class object}

\item{delete_zero_degree}{delete zero degree nodes}

\item{make_upper}{make these words uppercase}

\item{query_plot}{plot the column other than text}
}
\value{
biotext class object
}
\description{
process graph made from user-specified df
}
