% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{process_network_microbe}
\alias{process_network_microbe}
\title{process_network_microbe}
\usage{
process_network_microbe(
  ret,
  delete_zero_degree = TRUE,
  make_upper = NULL,
  disease_plot = FALSE,
  ec_plot = FALSE,
  metab = NULL,
  metab_col = NULL,
  metab_thresh = 0.2,
  mb_plot = FALSE,
  ec_file = NULL,
  up_tax_file = NULL
)
}
\arguments{
\item{ret}{biotext class object}

\item{delete_zero_degree}{exclude zero degree nodes from plotting}

\item{make_upper}{make these words to uppercase}

\item{disease_plot}{disease plotting}

\item{ec_plot}{enzyme plotting}

\item{mb_plot}{microbe plotting}

\item{ec_file}{enzyme database file}

\item{up_tax_file}{UniProt taxonomy file}

\item{metab_plot}{metabolite plotting}
}
\description{
perform tasks before plotting
like obtaining associated diseases to plot
}
