% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.R
\name{refseqDESeq2}
\alias{refseqDESeq2}
\title{refseqDESeq2}
\usage{
refseqDESeq2(res, condition, keyType = "ENSEMBL", argList = list())
}
\arguments{
\item{res}{results of DESeq2::results()}

\item{condition}{filtering condition}

\item{keyType}{key type of gene}

\item{argList}{argument list to pass to refseq()}
}
\value{
list of biotext class object
}
\description{
Return the biotext class object by specified filter in DESeq2 results object
}
