% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.R
\name{refseqWGCNA}
\alias{refseqWGCNA}
\title{refseqWGCNA}
\usage{
refseqWGCNA(wgcna, keyType = "ENSEMBL", argList = list())
}
\arguments{
\item{wgcna}{results of blockwiseModules()}

\item{keyType}{key type of gene}

\item{argList}{argument list to pass to refseq()}
}
\value{
list of biotext class object
}
\description{
Text mining WGCNA results
}
\details{
Return the list of biotext class object per cluster for the blockwise module results in WGCNA
}
\examples{
cols <- sample(c(1), 6, replace=TRUE)
names(cols) <- c("PNKP","ERCC2","ERCC3","IRF3","TP53","MCM6")
wg <- list("colors"=cols)
refseqWGCNA(wg, keyType="SYMBOL")
}
