% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{returnExample}
\alias{returnExample}
\title{returnExample}
\usage{
returnExample()
}
\value{
return example MEs and colors
}
\description{
Return an example dataset used in the analysis
}
\details{
Return an example dataset used in the analysis
}
\examples{
returnExample()
}
