% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textORA.R
\name{returnP}
\alias{returnP}
\title{returnP}
\usage{
returnP(name, matSorted, allFreqGeneSummary)
}
\arguments{
\item{name}{name of a gene}

\item{matSorted}{words frequency in the cluster}

\item{allFreqGeneSummary}{words frequency in the whole description}
}
\value{
p-values for the words
}
\description{
Calculate p-values
}
