% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{returnQuanteda}
\alias{returnQuanteda}
\title{returnQuanteda}
\usage{
returnQuanteda(
  ret,
  quantedaArgs,
  numWords,
  ngram,
  filterWords,
  additionalRemove,
  tfidf,
  collapse
)
}
\arguments{
\item{ret}{osplot object}

\item{quantedaArgs}{args to passed to tokens function}

\item{numWords}{number of words}

\item{ngram}{tokens_ngrams}

\item{filterWords}{filtered words}

\item{additionalRemove}{user-specified filter words}

\item{tfidf}{use dfm_tfidf or not}

\item{collapse}{collapse sentences}
}
\value{
osplot object after filtering using quanteda
}
\description{
simple function to generate dfm and export to wc* functions.
It did not utilize full functionality of quanteda, and 
will add better support.
}
