% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{returnSim}
\alias{returnSim}
\title{returnSim}
\usage{
returnSim(
  cllist,
  keyType = "ENTREZID",
  numLimit = 5000,
  target = "refseq",
  argList = list()
)
}
\arguments{
\item{cllist}{cluster list (named vector or list)}

\item{keyType}{keytype}

\item{numLimit}{threshold for gene number limit
default to 5000}

\item{target}{target to query}

\item{argList}{parameters to pass to refseq()}
}
\value{
similarity matrix
}
\description{
Return similarity matrix of cluster
based on intersection over union of words
}
\details{
The function accepts the list of character vectors
or named vector of identifiers to returns the textual similarity matrix.
}
\examples{
ex <- returnExample()
returnSim(ex$color, keyType="ENSEMBL")
}
