% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{set_filter_words}
\alias{set_filter_words}
\title{set_filter_words}
\usage{
set_filter_words(
  ret,
  exclude_by = "frequency",
  exclude_type = ">",
  exclude = "GS",
  exclude_number = 2000,
  filterMax = FALSE,
  additional_remove = NULL,
  pre = TRUE,
  pre_microbe = FALSE
)
}
\arguments{
\item{ret}{biotext object}

\item{exclude_by}{`frequency` or `tfidf`}

\item{exclude_type}{">" or "<"}

\item{exclude}{GS for using GeneSummary,}

\item{exclude_number}{threshold to exclude}

\item{filterMax}{filter based on pre-computed maximum values 
for each word}

\item{additional_remove}{your customized words to remove}

\item{pre}{remove pre-defined words}

\item{pre_microbe}{remove pre-defined words for microbes}
}
\value{
biotext class object
}
\description{
set filtering words to class
}
