% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{tag_words}
\alias{tag_words}
\title{tag_words}
\usage{
tag_words(
  ret,
  cl = FALSE,
  pvclAlpha = 0.95,
  whole = FALSE,
  num_words = 30,
  method = "correlation"
)
}
\arguments{
\item{ret}{biotext class object}

\item{cl}{cluster for parallel computing}

\item{pvclAlpha}{alpha value for `pvpick`}

\item{whole}{perform clustering on whole matrix (take time)}

\item{num_words}{if set, subset to this number of words}

\item{method}{method.dist argument in pvclust}
}
\value{
biotext class object
}
\description{
tag the words based on pvclust and word matrix
}
\examples{
obtain_refseq(c("IRF3","PNKP")) |> 
make_corpus() |> make_TDM() |> tag_words()
}
