% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.R
\name{TextMarkers}
\alias{TextMarkers}
\title{TextMarkers}
\usage{
TextMarkers(
  df,
  keyType = "SYMBOL",
  type = "wc",
  genePlot = TRUE,
  genePlotNum = 5,
  colorText = TRUE,
  wcArg = NULL,
  raw = FALSE,
  col = NULL
)
}
\arguments{
\item{df}{result of FindAllMarkers()}

\item{keyType}{keytype}

\item{type}{wc or network}

\item{genePlot}{whether to plot relevant genes}

\item{genePlotNum}{number of genes to plot}

\item{colorText}{colorlize text or not, default to TRUE}

\item{wcArg}{parameters to pass to ggwordcloud}

\item{raw}{obtain raw results of wcGeneSummary instead of plot}

\item{col}{color to be used in wordcloud}
}
\description{
Take results of Seurat::FindAllMarkers as input,
and plot wordcloud or network for all the clusters.
}
