% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textORA.R
\name{textORA}
\alias{textORA}
\title{textORA}
\usage{
textORA(queries, notGene = FALSE, bg = NULL)
}
\arguments{
\item{queries}{gene list (Entrez ID)}

\item{notGene}{boolean specifying type of input, default to gene IDs}

\item{bg}{default to NULL}
}
\value{
p-values for the words

p-values for the words
}
\description{
Return p-values from hypergeometric distribution
(experimental)
}
\examples{
textORA(c("2067","2068","2071","2072"))
}
