% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcAbst.R
\name{wcAbst}
\alias{wcAbst}
\title{wcAbst}
\usage{
wcAbst(
  queries,
  redo = NA,
  madeUpper = c("dna", "rna"),
  palette = c("blue", "red"),
  numWords = 30,
  scaleRange = c(5, 10),
  showLegend = FALSE,
  plotType = "wc",
  colorText = FALSE,
  corThresh = 0.6,
  layout = "nicely",
  edgeLabel = FALSE,
  edgeLink = TRUE,
  ngram = NA,
  deleteZeroDeg = TRUE,
  additionalRemove = NA,
  ...
)
}
\arguments{
\item{queries}{gene symbols (max: 5)}

\item{redo}{if plot in other parameters, input the previous list}

\item{madeUpper}{make the words uppercase in resulting plot}

\item{palette}{palette for color gradient in correlation network}

\item{numWords}{the number of words to be shown}

\item{scaleRange}{scale for label and node size in correlation network}

\item{showLegend}{whether to show legend in correlation network}

\item{plotType}{"wc" or "network"}

\item{colorText}{color text label based on frequency in correlation network}

\item{corThresh}{the correlation threshold}

\item{layout}{the layout for correlation network, defaul to "nicely"}

\item{edgeLabel}{if TRUE, plot the edge label (default: FALSE)}

\item{edgeLink}{if FALSE, use geom_edge_diagonal}

\item{ngram}{default to NA (1)}

\item{deleteZeroDeg}{delete zero degree node from plot in correlation network}

\item{additionalRemove}{specific words to be excluded}

\item{...}{parameters to pass to wordcloud()}
}
\value{
list of data frame and ggplot2 object
}
\description{
make word cloud or correlation network from PubMed abstracts using easyPubMed
}
\examples{
wcAbst("DDX41")
}
