% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcBSDB.R
\name{wcBSDB}
\alias{wcBSDB}
\title{wcBSDB}
\usage{
wcBSDB(
  mbList,
  excludeFreq = 1000,
  excludeTfIdf = NA,
  additionalRemove = NA,
  target = "title",
  apiKey = NULL,
  pre = FALSE,
  madeUpper = c("dna", "rna"),
  redo = NA,
  pal = c("blue", "red"),
  numWords = 15,
  scaleRange = c(5, 10),
  showLegend = FALSE,
  edgeLabel = FALSE,
  mbPlot = FALSE,
  ngram = NA,
  plotType = "wc",
  disPlot = FALSE,
  colorText = FALSE,
  corThresh = 0.6,
  tag = FALSE,
  layout = "nicely",
  edgeLink = TRUE,
  deleteZeroDeg = TRUE,
  ...
)
}
\arguments{
\item{mbList}{microbe list}

\item{excludeFreq}{exclude words with overall frequency above excludeFreq
default to 0}

\item{excludeTfIdf}{exclude based on tfidf (default: NA)}

\item{additionalRemove}{specific words to be excluded}

\item{target}{"title" or "abstract"}

\item{apiKey}{api key for eutilities}

\item{pre}{predefined filter words}

\item{madeUpper}{make the words uppercase in resulting plot}

\item{redo}{if "abstract" is chosen in target, one can provide resulting object again}

\item{pal}{palette for color gradient in correlation network and tag coloring}

\item{numWords}{the number of words to be shown}

\item{scaleRange}{scale for label and node size in correlation network}

\item{showLegend}{whether to show legend in correlation network}

\item{edgeLabel}{if TRUE, plot the edge label (default: FALSE)}

\item{mbPlot}{plot microbe names}

\item{ngram}{default to NA (1)}

\item{plotType}{"wc" or "network"}

\item{disPlot}{plot diseases}

\item{colorText}{color text label based on frequency in correlation network}

\item{corThresh}{the correlation threshold}

\item{tag}{perform pvclust on words and colorlize them in wordcloud}

\item{layout}{the layout for correlation network, defaul to "nicely"}

\item{edgeLink}{if FALSE, use geom_edge_diagonal}

\item{deleteZeroDeg}{delete zero degree node from plot in correlation network}

\item{...}{parameters to pass to wordcloud()}
}
\value{
list of data frame and ggplot2 object
}
\description{
Plot wordcloud of BugSigDB
}
\examples{
mbList <- c("Veillonella dispar")
wcBSDB(mbList)
}
