% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcEC.R
\name{wcEC}
\alias{wcEC}
\title{wcEC}
\usage{
wcEC(file, ecnum, onlyTerm = FALSE, onlyDf = FALSE, ...)
}
\arguments{
\item{file}{file downloaded from expasy}

\item{ecnum}{candidate ecnum, like those obtained from eggNOG-mapper}

\item{onlyTerm}{only return quoted queries to wcAbst}

\item{onlyDf}{only return ec description data.frame
if onlyTerm and onlyDf are both specified, onlyTerm have priority}

\item{...}{passed to osplot(target="pubmed")}
}
\description{
Query the Enzyme Comission number and obtain description,
and search pubmed for these enzymes and make word cloud and
correlation network. Need "enzyme.dat" from ExPASy
(https://enzyme.expasy.org/).
}
