% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcGeneSummary.R
\name{wcGeneSummary}
\alias{wcGeneSummary}
\title{wcGeneSummary}
\usage{
wcGeneSummary(
  geneList,
  excludeFreq = 5000,
  additionalRemove = NA,
  madeUpper = c("dna", "rna"),
  organism = 9606,
  ...
)
}
\arguments{
\item{geneList}{ENTREZID list}

\item{excludeFreq}{exclude words with overall frequency above excludeFreq, default to 5000}

\item{additionalRemove}{NA}

\item{madeUpper}{make the words uppercase in resulting plot}
}
\value{
list of data frame and ggplot2 object
}
\description{
Plot wordcloud of RefSeq description obtained by GeneSummary
}
\examples{
wcGeneSummary(geneList)
}
