% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{changeLayout}
\alias{changeLayout}
\title{changeLayout}
\usage{
changeLayout(g, layout_func)
}
\arguments{
\item{g}{biotext object}

\item{layout_func}{layout function in igraph}
}
\value{
biotext class object
}
\description{
Change the layout of biotext object and plot.
}
\details{
Change the layout of biotext object and plot.
}
\examples{
refseq(c("IRF3","PNKP","DDX41","ERCC1","ERCC2","XRCC1")) |> changeLayout(igraph::layout_nicely)
}
