% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{changeLayout}
\alias{changeLayout}
\title{changeLayout}
\usage{
changeLayout(g, layout_func)
}
\arguments{
\item{g}{biotext object}

\item{layout_func}{layout function in igraph}
}
\value{
biotext class object
}
\description{
changeLayout
}
\examples{
wcGeneSummary(c("IRF3","PNKP","DDX41")) |> changeLayout(igraph::layout_nicely)
}
