% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareWordNet.R
\name{compareWordNet}
\alias{compareWordNet}
\title{compareWordNet}
\usage{
compareWordNet(
  wc1,
  wc2,
  titles = c("title1", "title2"),
  layout = "nicely",
  hull = FALSE,
  size = 4,
  tag = FALSE
)
}
\arguments{
\item{wc1}{results of wc functions (plotType="network")}

\item{wc2}{results of wc functions (plotType="network")}

\item{titles}{title to be shown on plot}

\item{layout}{layout}

\item{hull}{show category by hull}

\item{size}{node size}

\item{tag}{show tag on plot}
}
\value{
plot comparing two gene clusters
}
\description{
compare two gene clusters based on words
}
\examples{
wc1 <- wcGeneSummary("DDX41", plotType="network")
wc2 <- wcGeneSummary("IRF3", plotType="network")
compare <- compareWordNet(wc1, wc2)
}
