% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareWordNet.R
\name{compareWordNet}
\alias{compareWordNet}
\title{compareWordNet}
\usage{
compareWordNet(
  listOfNets,
  titles = NULL,
  layout = "nicely",
  hull = FALSE,
  size = "freq",
  conc = 1,
  tag = FALSE,
  tagLevel = 1,
  edgeLink = TRUE,
  freqMean = FALSE,
  scaleRange = c(5, 10),
  ovlThresh = 0,
  returnNet = FALSE,
  colPal = "Pastel1",
  colNum = 5,
  colorText = FALSE,
  community = FALSE,
  returnClass = TRUE
)
}
\arguments{
\item{listOfNets}{list consisting results of wc* functions (plotType="network")}

\item{titles}{title to be shown on plot}

\item{layout}{layout}

\item{hull}{show category by hull}

\item{size}{node size, freq, overlap or numeric number}

\item{conc}{concavity parameter}

\item{tag}{show tag on plot}

\item{tagLevel}{words that tagged in this many networks will be included}

\item{edgeLink}{whether to use link or diagonal}

\item{freqMean}{how to concatenate frequency (TRUE: mean, FALSE: sum)}

\item{scaleRange}{point size scaling}

\item{ovlThresh}{show text with this number of overlap between graphs}

\item{returnNet}{return only the network (ig)}

\item{colPal}{color palette to be used in RColorBrewer}

\item{colNum}{color number to be used in plot}

\item{colorText}{whether to color text based on category}

\item{community}{compare based on community (igraph), override tag}

\item{returnClass}{return biotext class object, default to TRUE}
}
\value{
plot comparing gene clusters
}
\description{
compare two gene clusters based on words
}
\examples{
net1 <- refseq(c("DDX41","IRF3"), plotType="network")
net2 <- refseq(c("DDX41","PNKP"), plotType="network")
compare <- compareWordNet(list(net1, net2))
}
