% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{connectGenes}
\alias{connectGenes}
\title{connectGenes}
\usage{
connectGenes(nets, query_word, return_tbl_graph = FALSE, neighbors = FALSE)
}
\arguments{
\item{nets}{named list of nets (if no names are given, automatically set)}

\item{query_word}{word to connect}

\item{return_tbl_graph}{return tbl_graph}

\item{neighbors}{obtain neighbors words for queried words, default to FALSE}
}
\value{
igraph object
}
\description{
When an interesting word is found in two or more networks,
connect the words and gene names and return the graph.
Note that genePlot must be set to TRUE in all the networks.
}
\examples{
n1 <- refseq(c("IRF3","PNKP","DDX41","ERCC1","ERCC2","XRCC1"), genePlot=TRUE) 
n2 <- refseq(c("IRF3","PNKP","DDX41","ERCC1","ERCC2","XRCC1"), genePlot=TRUE)
connectGenes(list(n1, n2), query_word="dna")
}
