% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exportCyjs}
\alias{exportCyjs}
\title{exportCyjs}
\usage{
exportCyjs(g, rootDir, netDir)
}
\arguments{
\item{g}{igraph object}

\item{rootDir}{root directory path}

\item{netDir}{directory to store scripts}
}
\description{
Export Cytoscape.js script, HTML and stylesheet for the graph and image
}
