% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exportVisjs}
\alias{exportVisjs}
\title{exportVisjs}
\usage{
exportVisjs(g, rootDir, netDir)
}
\arguments{
\item{g}{igraph object}

\item{rootDir}{root directory path}

\item{netDir}{directory to store scripts}
}
\value{
return nothing, export to a specified directory
}
\description{
Export vis.js script, HTML and stylesheet for the graph and image
}
\examples{
library(igraph)
g <- graph_from_literal( ME1-+ME2 )
V(g)$image <- c("path1","path2")
V(g)$shape <- c("image","image")
V(g)$size <- c(1,1)
\donttest{exportVisjs(g, "./", "net")}
}
