% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportWCNetwork.R
\name{exportWCNetwork}
\alias{exportWCNetwork}
\title{exportWCNetwork}
\usage{
exportWCNetwork(
  g,
  geneList,
  dir = "network",
  colors = NULL,
  scaleMin = 50,
  scaleMax = 200,
  wcScale = 20,
  keyType = "SYMBOL",
  orgDb = org.Hs.eg.db,
  wcArgs = list(),
  numWords = 50
)
}
\arguments{
\item{g}{igraph (node name corresponding to gene cluster name)}

\item{geneList}{named list of gene IDs}

\item{dir}{directory to output the files, default to current directory}

\item{colors}{color palettes to use for each cluster
named list, or random palette will be assigned if NULL}

\item{scaleMin}{minimum scale size for nodes}

\item{scaleMax}{maximum scale size for nodes}

\item{wcScale}{scaling size for wordcloud}

\item{keyType}{type of gene IDs (default to SYMBOL)}

\item{orgDb}{annotation database}

\item{wcArgs}{parameters for ggwordcloud}

\item{numWords}{number of words}
}
\value{
export the Cytoscape.js network
}
\description{
export the wordcloud network provided the gene cluster network
Node size will be the number of genes in the cluster
}
