% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findTerm}
\alias{findTerm}
\title{findTerm}
\usage{
findTerm(
  query,
  listOfGenes,
  split = FALSE,
  ngram = 1,
  tfidf = TRUE,
  calc = "sum",
  keyType = "SYMBOL",
  argList = list()
)
}
\arguments{
\item{query}{query words}

\item{listOfGenes}{named list of genes}

\item{split}{split the query by space, default to FALSE}

\item{ngram}{use ngram or not}

\item{tfidf}{use tfidf when creating TDM or not}

\item{calc}{"sum", "mean" or "highest"}

\item{keyType}{key type of listOfGenes}

\item{argList}{passed to refseq}
}
\value{
named list of frequency
}
\description{
Find queried terms in list of gene clusters and return frequency
}
\details{
Find queried terms in list of gene clusters and return frequency
}
\examples{
query <- "DNA repair"
lg <- list()
lg[["sample"]] <- c("ERCC1","ERCC2")
findTerm(query, lg)

}
