% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotextgraph.R
\name{getSlot}
\alias{getSlot}
\title{getSlot}
\usage{
getSlot(x, slot)
}
\arguments{
\item{x}{biotext object}

\item{slot}{slot name}
}
\value{
attribute value
}
\description{
get the slot value from biotext object
}
\examples{
n1 <- refseq(c("IRF3","PNKP","DDX41","ERCC1","ERCC2","XRCC1")) 
getSlot(n1, "igraphRaw")
}
