% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getUPtax}
\alias{getUPtax}
\title{getUPtax}
\usage{
getUPtax(file, candUP, candTax = NULL)
}
\arguments{
\item{file}{downloaded file}

\item{candUP}{candidate UniProt organism identification codes}

\item{candTax}{candidate taxonomy name}
}
\value{
data.frame consisting of taxonomy name and UniProt IDs
}
\description{
Obtain the list of UniProt organism identification codes, by querying taxonomy or UniProt codes.
https://ftp.uniprot.org/pub/databases/uniprot/current_release/knowledgebase/complete/docs/speclist.txt
The file above must be downloaded and specified to file argument.
}
\examples{
file <- "speclist.txt"
\dontrun{getUPtax(file, candUP="all")}
}
