% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWordsOnDendro.R
\name{getWordsOnDendro}
\alias{getWordsOnDendro}
\title{getWordsOnDendro}
\usage{
getWordsOnDendro(
  dhc,
  geneVec,
  geneNumLimit = 1000,
  geneVecType = "ENSEMBL",
  numberOfWords = 25,
  excludeFreq = 10000,
  orgDb = org.Hs.eg.db,
  madeUpper = c("rna", "dna")
)
}
\arguments{
\item{dhc}{dendrogram}

\item{geneVec}{gene-named vector of node names in dendrogram}

\item{geneNumLimit}{when the gene number is above this threshold, pyramid plots are not produced}

\item{geneVecType}{type of the name of geneVec (default: ENSEMBL)}

\item{numberOfWords}{the number of words to plot (default: 25)}

\item{excludeFreq}{words with the frequency above this threshold is excluded beforehand}

\item{orgDb}{organism database}

\item{madeUpper}{words with uppercase}
}
\value{
list of pyramid plot grobs and its positions
}
\description{
Get grobs to plot on the dendrogram with the position information
}
\examples{
\dontrun{getWordsOnDendro(dhc, geneVec)}
}
