% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWordsOnDendro.R
\name{getWordsOnDendro}
\alias{getWordsOnDendro}
\title{getWordsOnDendro}
\usage{
getWordsOnDendro(
  dhc,
  geneVec,
  geneNumLimit = 1000,
  geneVecType = "ENSEMBL",
  useWC = FALSE,
  useRandomColor = FALSE,
  numberOfWords = 25,
  showType = "ID",
  wcArgs = list(),
  highlight = NULL,
  textSize = 3.5,
  wcScale = 3,
  wrap = NULL,
  candidateNodes = NULL,
  takeIntersect = TRUE,
  useDf = NULL,
  bg.colour = NULL,
  type = "words",
  argList = list(),
  useFunc = NULL,
  useggfx = NULL,
  normalizeByClusterNum = TRUE,
  autoSize = TRUE
)
}
\arguments{
\item{dhc}{dendrogram}

\item{geneVec}{gene-named vector of node names in dendrogram}

\item{geneNumLimit}{when the gene number is above this threshold,
pyramid plots are not produced}

\item{geneVecType}{type of the name of geneVec (default: ENSEMBL)}

\item{useWC}{use wordcloud}

\item{useRandomColor}{use random colors on wordclouds}

\item{numberOfWords}{the number of words to plot (default: 25)}

\item{showType}{when "enrich", which labels to show}

\item{wcArgs}{argument list for ggwordcloud}

\item{highlight}{words to highlight}

\item{textSize}{text size in pyramid plots}

\item{wcScale}{max_size of wordcloud}

\item{wrap}{wrap the strings}

\item{candidateNodes}{if NULL, all nodes are investigated}

\item{takeIntersect}{take intersection or frequent words}

\item{useDf}{data.frame to subset when manual function is specified}

\item{bg.colour}{background color for wordcloud}

\item{type}{"words" or "enrich"}

\item{argList}{passed to refseq}

\item{useFunc}{function to summarize text}

\item{useggfx}{use ggfx on resulting plot}

\item{normalizeByClusterNum}{normalize frequency by ID numbers or not}

\item{autoSize}{size the text based on grob width}
}
\value{
list of pyramid plot grobs and its positions
}
\description{
Get grobs to plot on the dendrogram with the position information
}
\examples{
mod <- returnExample()
result <- hclust(dist(t(mod$MEs)))
dhc <- result |>
    as.dendrogram() |>
    dendextend::hang.dendrogram()
geneVec <- paste0("ME", mod$colors)
names(geneVec) <- names(mod$colors)
getWordsOnDendro(dhc, geneVec, numberOfWords = 2)
}
