% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_sc_wordcloud.R
\name{ggplot_add.geom_sc_wordcloud}
\alias{ggplot_add.geom_sc_wordcloud}
\title{ggplot_add.geom_sc_wordcloud
use ggplot_add to populate single-cell plot with textual information}
\usage{
\method{ggplot_add}{geom_sc_wordcloud}(object, plot, object_name)
}
\arguments{
\item{object}{An object to add to the plot}

\item{plot}{The ggplot object to add object to}

\item{object_name}{The name of the object to add}
}
\value{
ggplot
}
\description{
ggplot_add.geom_sc_wordcloud
use ggplot_add to populate single-cell plot with textual information
}
