% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{graph_cluster}
\alias{graph_cluster}
\title{graph_cluster}
\usage{
graph_cluster(ret, func = igraph::cluster_leiden, factorize = TRUE)
}
\arguments{
\item{ret}{biotext class object}

\item{func}{community detection algorithm in igraph}

\item{factorize}{convert to factor upon assigning}
}
\value{
biotext class object
}
\description{
graph_cluster
}
