% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeCorpus}
\alias{makeCorpus}
\title{makeCorpus}
\usage{
makeCorpus(docs, filterWords, additionalRemove, numOnly, stem, lower = TRUE)
}
\arguments{
\item{docs}{corpus to clean}

\item{filterWords}{words to filter based on frequency}

\item{additionalRemove}{words to filter}

\item{numOnly}{delete number only}

\item{stem}{use stem or not}

\item{lower}{transform lower for corpus}
}
\value{
cleaned corpus
}
\description{
makeCorpus
}
