% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_sc_wordcloud.R
\name{obtainMarkersWC}
\alias{obtainMarkersWC}
\title{obtainMarkersWC}
\usage{
obtainMarkersWC(
  markers,
  cols,
  wcArgs,
  eps = 1e-10,
  sortBy = "avg_log2FC",
  decreasing = TRUE,
  scaleNumber = 10,
  wcScale = 5,
  geneNum = 50,
  withggfx = NULL,
  ggfxParams = list()
)
}
\arguments{
\item{markers}{marker data frame}

\item{cols}{list of colors}

\item{wcArgs}{arguments for ggwordcloud}

\item{eps}{when taking log of p-values, this value will be added}

\item{sortBy}{default to avg_log2FC, "log10p" can be specified.}

\item{decreasing}{sort by decreasing order or not}

\item{scaleNumber}{scale the frequency by this number}

\item{wcScale}{scale for scale_size_area()}

\item{geneNum}{number of genes to be included in wordclouds}

\item{withggfx}{applying ggfx filters}

\item{ggfxParams}{parameter list for ggfx}
}
\value{
list of ggplot containing wordclouds
}
\description{
obtain wordclouds of cluster markers identified in Seurat
}
\details{
Using the results of marker gene identification such as `FindAllMarkers` from Seurat,
Recursively summarize the textual information of markers and output the wordclouds.
}
\examples{
markers <- data.frame(p_val=c(0.01, 0.01),gene=c("PNKP","DDX41"),cluster=c("1","1"))
colors <- list("1"="red")
obtainMarkersWC(markers, sortBy="p_val", cols=colors, wcArgs=list(), geneNum=2)
}
