% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_sc_wordcloud.R
\name{obtainMarkersWCScran}
\alias{obtainMarkersWCScran}
\title{obtainMarkersWCScran
make gene wordcloud from scran::findMarkers() results}
\usage{
obtainMarkersWCScran(
  markers,
  cols,
  wcArgs,
  eps = 1e-10,
  sortBy = "summary.logFC",
  decreasing = TRUE,
  scaleNumber = 10,
  wcScale = 5,
  geneNum = 50,
  withggfx = NULL,
  ggfxParams = list()
)
}
\arguments{
\item{markers}{marker list}

\item{cols}{list of colors}

\item{wcArgs}{arguments for ggwordcloud}

\item{eps}{when taking log of p-values, this value will be added}

\item{sortBy}{default to summary.logFC, "log10p" can be specified.}

\item{decreasing}{sort by decreasing order or not}

\item{scaleNumber}{scale the frequency by this number}

\item{wcScale}{scale for scale_size_area()}

\item{geneNum}{number of genes to be included in wordclouds}
}
\description{
obtainMarkersWCScran
make gene wordcloud from scran::findMarkers() results
}
