% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtainTextPosition}
\alias{obtainTextPosition}
\title{obtainTextPosition}
\usage{
obtainTextPosition(ig, sort.by = node_rank_fabric(), verbose = FALSE)
}
\arguments{
\item{ig}{igraph}

\item{sort.by}{the argument to be passed to fabric layout function}

\item{verbose}{show logs}
}
\value{
tbl_graph
}
\description{
Obtain text position in biofabric layout
}
\details{
The function obtains the horizontal text position 
for the plot in the `biofabric` layout implemented in ggraph.
}
\examples{
testgenes <- c("IRF3","PNKP","DDX41","ERCC1","ERCC2","XRCC1")
getSlot(refseq(testgenes), "igraph") |> obtainTextPosition()
}
