% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcAbst.R
\name{ospa}
\alias{ospa}
\title{ospa}
\usage{
ospa(
  queries,
  redo = NULL,
  madeUpper = c("dna", "rna"),
  target = "abstract",
  usefil = NA,
  filnum = 0,
  sortOrder = "relevance",
  pvclAlpha = 0.95,
  numOnly = TRUE,
  delim = "OR",
  limit = 10,
  geneUpper = TRUE,
  apiKey = NULL,
  tfidf = FALSE,
  cl = FALSE,
  pal = c("blue", "red"),
  numWords = 30,
  scaleRange = c(5, 10),
  showLegend = FALSE,
  plotType = "wc",
  colorText = FALSE,
  quote = FALSE,
  corThresh = 0.2,
  layout = "nicely",
  tag = FALSE,
  tagWhole = FALSE,
  onlyCorpus = FALSE,
  onlyTDM = FALSE,
  bn = FALSE,
  R = 20,
  retMax = 10,
  edgeLabel = FALSE,
  edgeLink = TRUE,
  ngram = NA,
  genePlot = FALSE,
  onlyDf = FALSE,
  nodePal = palette(),
  deleteZeroDeg = TRUE,
  additionalRemove = NA,
  orgDb = org.Hs.eg.db,
  preset = FALSE,
  onWholeDTM = FALSE,
  madeUpperGenes = TRUE,
  stem = FALSE,
  ...
)
}
\description{
alias for wcAbst
}
\examples{
ospa("DDX41")
}
