% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osplot.R
\name{osplot}
\alias{osplot}
\title{osplot}
\usage{
osplot(target, argList)
}
\arguments{
\item{target}{"pubmed", "bugsigdb", "refseq", "ec"}

\item{argList}{passed to each function}
}
\value{
list of data frames and ggplot2 object
}
\description{
wrapper for functions wcGeneSummary, wcAbst, and wcBSDB
}
\examples{
geneList <- c("DDX41")
osplot("refseq", argList=list(geneList=geneList))
}
