% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parseMetaCycPathway}
\alias{parseMetaCycPathway}
\title{parseMetaCycPathway}
\usage{
parseMetaCycPathway(file, candSp, withTax = FALSE)
}
\arguments{
\item{file}{path to pathways.dat}

\item{candSp}{species to grepl}

\item{withTax}{parse taxonomy information}
}
\description{
parse MetaCyc "pathways.dat"
}
