% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parseMetaCycPathway}
\alias{parseMetaCycPathway}
\title{parseMetaCycPathway}
\usage{
parseMetaCycPathway(
  file,
  candSp,
  withTax = FALSE,
  noComma = FALSE,
  clear = FALSE
)
}
\arguments{
\item{file}{path to pathways.dat}

\item{candSp}{species to grepl}

\item{withTax}{parse taxonomy information}

\item{noComma}{no comma separated when taxonomy parsing}

\item{clear}{delete HTML tags and some symbols}
}
\value{
data.frame of MetaCyc pathway information
}
\description{
parse MetaCyc "pathways.dat"
}
\examples{
file <- "pathways.dat"
\dontrun{parseMetaCycPathway(file, candSp="all")}
}
