% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathviewText.R
\name{pathviewText}
\alias{pathviewText}
\title{pathviewText}
\usage{
pathviewText(
  geneList,
  keyType,
  pid,
  pal = "RdBu",
  target = "refseq",
  numWords = 20,
  ...
)
}
\arguments{
\item{geneList}{gene ID list}

\item{keyType}{keytype for queried gene}

\item{pid}{pathway id (hsa)}

\item{pal}{palette to be passed to RColorBrewer
for coloring of nodes}

\item{target}{target DB for word}

\item{numWords}{number of words in barplot}

\item{...}{passed to wc functions}
}
\description{
colorize pathview map, and link to word frequency
}
