% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareWordNet.R
\name{plotDynamic}
\alias{plotDynamic}
\title{plotDynamic}
\usage{
plotDynamic(listOfNets, alpha = 0.2, titles = NULL, tag = FALSE)
}
\arguments{
\item{listOfNets}{list consisting results of wc* functions (plotType="network")}

\item{alpha}{pass to layout_as_dynamic}

\item{titles}{title to be shown on plot}

\item{tag}{show tag on plot}
}
\value{
plot comparing gene clusters
}
\description{
list network of words using graphlayouts::layout_as_dynamic
}
\examples{
wc1 <- wcGeneSummary("DDX41", plotType="network")
wc2 <- wcGeneSummary("IRF3", plotType="network")
compare <- plotDynamic(list(wc1, wc2))
}
