% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWordsOnDendro.R
\name{plotEigengeneNetworksWithWords}
\alias{plotEigengeneNetworksWithWords}
\title{plotEigengeneNetworksWithWords}
\usage{
plotEigengeneNetworksWithWords(
  MEs,
  colors,
  nboot = 100,
  numberOfWords = 10,
  geneNumLimit = 1000,
  geneVecType = "ENSEMBL",
  useWC = FALSE,
  border = TRUE,
  type = "words",
  wcScale = 3,
  candidateNodes = NULL,
  takeIntersect = TRUE,
  showType = "ID",
  textSize = 3.5,
  horiz = FALSE,
  dendPlot = "pvclust",
  dhc = NULL,
  wcArg = list(),
  useDf = NULL,
  useWGCNA = TRUE,
  highlight = NULL,
  argList = list(),
  useFunc = NULL
)
}
\arguments{
\item{MEs}{module eigengene data (data.frame of row as sample and col as gene cluster IDs)}

\item{colors}{named vector of cluster}

\item{nboot}{pvclust bootstrap number}

\item{numberOfWords}{number of words to be included in pyramid plots}

\item{geneNumLimit}{clusters with gene number above this threshold are ignored}

\item{geneVecType}{type of IDs in `colors`}

\item{useWC}{plot wordcloud instead of pyramid plot}

\item{border}{whether to draw border on pyramid plots}

\item{type}{"words" or "enrich"}

\item{wcScale}{max_size of wordcloud}

\item{candidateNodes}{if NULL, all nodes are investigated}

\item{takeIntersect}{take intersection or frequent words}

\item{showType}{when "enrich", which labels to show}

\item{textSize}{text size in pyramid plots}

\item{horiz}{horizontal plot or not}

\item{dendPlot}{type of dendrogram plot}

\item{dhc}{user-specified dendrogram}

\item{wcArg}{argument list, pass to ggwordcloud}

\item{useDf}{data.frame to subset when manual function is specified
"query" column will be used to subset}

\item{useWGCNA}{pad named color vector with prefix "ME"}

\item{highlight}{words to highlight}

\item{argList}{passed to wcGeneSummary()}

\item{useFunc}{if not specified, use RefSeq summary data}
}
\value{
plot of dendrogram between module eigengenes with word annotations
}
\description{
plot an eigengene dendrogram with word annotation
}
\examples{
mod <- returnExample()
plotEigengeneNetworksWithWords(mod$MEs, mod$colors)
}
