% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWordsOnDendro.R
\name{plotEigengeneNetworksWithWords}
\alias{plotEigengeneNetworksWithWords}
\title{plotEigengeneNetworksWithWords}
\usage{
plotEigengeneNetworksWithWords(
  MEs,
  colors,
  nboot = 100,
  numberOfWords = 10,
  geneNumLimit = 1000,
  geneVecType = "ENSEMBL",
  useWC = FALSE,
  border = TRUE,
  type = "words",
  wcScale = 3,
  candidateNodes = NULL,
  takeIntersect = TRUE,
  showType = "ID",
  textSize = 3.5,
  horiz = FALSE,
  dendPlot = "pvclust",
  dhc = NULL,
  wcArgs = list(),
  useDf = NULL,
  useWGCNA = TRUE,
  spacer = 0.005,
  wrap = NULL,
  highlight = NULL,
  argList = list(),
  useFunc = NULL,
  returnGlobOnly = FALSE,
  tipWC = FALSE,
  tipWCNodes = NULL,
  imageDir = NULL,
  wh = 5,
  al = TRUE,
  offset = 0.2,
  tipSize = 0.3,
  asp = 1.5,
  horizontalSpacer = 0,
  useRandomColor = FALSE,
  bg.colour = NULL,
  useggfx = NULL,
  ggfxParams = list()
)
}
\arguments{
\item{MEs}{module eigengene data (data.frame of row as sample and col as gene cluster IDs)}

\item{colors}{named vector of cluster}

\item{nboot}{pvclust bootstrap number}

\item{numberOfWords}{number of words to be included in pyramid plots}

\item{geneNumLimit}{clusters with gene number above this threshold are ignored}

\item{geneVecType}{type of IDs in `colors`}

\item{useWC}{plot wordcloud instead of pyramid plot}

\item{border}{whether to draw border on pyramid plots}

\item{type}{"words" or "enrich"}

\item{wcScale}{max_size of wordcloud}

\item{candidateNodes}{if NULL, all nodes are investigated}

\item{takeIntersect}{take intersection or frequent words}

\item{showType}{when "enrich", which labels to show}

\item{textSize}{text size in pyramid plots}

\item{horiz}{horizontal plot or not}

\item{dendPlot}{type of dendrogram plot
if dendPlot=="ggtree", provide those accepted by ggtree function to `dhc`.}

\item{dhc}{user-specified dendrogram}

\item{wcArgs}{argument list, pass to ggwordcloud}

\item{useDf}{data.frame to subset when manual function is specified
"query" column will be used to subset}

\item{useWGCNA}{pad named color vector with prefix "ME"}

\item{spacer}{spacing for grob}

\item{wrap}{wrap string}

\item{highlight}{words to highlight}

\item{argList}{passed to wcGeneSummary()}

\item{useFunc}{if not specified, use RefSeq summary data}

\item{returnGlobOnly}{default to FALSE}

\item{tipWC}{using ggtree, show WC on tip}

\item{tipWCNodes}{node names to plot}

\item{imageDir}{image directory to output WC}

\item{wh}{width and height of wordcloud}

\item{al}{positional argument to passed to ggtree::geom_tiplab (ggimage::geom_image)}

\item{offset}{positional argument to passed to ggtree::geom_tiplab (ggimage::geom_image)}

\item{tipSize}{positional argument to passed to ggtree::geom_tiplab (ggimage::geom_image)}

\item{asp}{aspect ratio for ggimage::geom_image}

\item{horizontalSpacer}{horizontal spacer for annotation}

\item{useRandomColor}{use random color on wordclouds}

\item{bg.colour}{use shadowtext for wordcloud, override border to FALSE}

\item{useggfx}{filter in ggfx to apply to wordcloud, default to NULL}

\item{ggfxParams}{parameters to pass to ggfx geom}

\item{normalizeByClusterNum}{normalize frequency by ID numbers or not}
}
\value{
plot of dendrogram between module eigengenes with word annotations
}
\description{
plot an eigengene dendrogram with word annotation
}
\examples{
mod <- returnExample()
plotEigengeneNetworksWithWords(mod$MEs, mod$colors)
}
