% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotextgraph.R
\name{plotNet}
\alias{plotNet}
\title{plotNet}
\usage{
plotNet(
  x,
  layout = "nicely",
  edgeLink = FALSE,
  edgeLabel = FALSE,
  showLegend = FALSE,
  fontFamily = "sans",
  tagPalette = NULL,
  catColors = NULL,
  queryColor = "grey",
  pal = c("blue", "red"),
  colorize = FALSE,
  discreteColorWord = FALSE,
  useSeed = 42,
  autoScale = FALSE,
  scaleRange = c(5, 10),
  scaleEdgeWidth = c(1, 3),
  naEdgeColor = "grey",
  colorText = FALSE,
  asis = FALSE
)
}
\arguments{
\item{layout}{layout specification}

\item{asis}{plot the original network (default to FALSE)}
}
\value{
network visualization
}
\description{
plot the network changing the visualization parameters
}
