% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotextgraph.R
\name{plotORA}
\alias{plotORA}
\title{plotORA}
\usage{
plotORA(x, thresh = 0.001)
}
\arguments{
\item{x}{biotext object}

\item{thresh}{hline to draw in plot}
}
\value{
volcano plot for ORA results
}
\description{
plot volcano-plot like plot for ORA results
}
\examples{
net1 <- refseq(c("DDX41","IRF3"), ora=TRUE)
plotORA(net1)
}
