% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.R
\name{plotReducedDimWithTexts}
\alias{plotReducedDimWithTexts}
\title{plotReducedDimWithTexts}
\usage{
plotReducedDimWithTexts(
  sce,
  marker.info,
  colour_by = "label",
  point_alpha = 0.4,
  use_shadowtext = TRUE,
  bg.colour = "white",
  which.label = NULL,
  wc_alpha = 1,
  wcScale = 5,
  rot.per = 0.4,
  rad = NULL,
  top = 5,
  gene_name = FALSE,
  withggfx = NULL,
  ggfxParams = list(),
  sortBy = "summary.logFC",
  scaleNumber = 2,
  decreasing = TRUE,
  geneNum = 50,
  random.order = FALSE,
  dimred = "PCA",
  base_ellipse = FALSE,
  base_dens = FALSE,
  withTitle = FALSE,
  args = list()
)
}
\arguments{
\item{sce}{sce object}

\item{marker.info}{results of findMarkers()}

\item{colour_by}{colorize by this label}

\item{point_alpha}{point alpha}

\item{use_shadowtext}{use shadowtext for wordcloud}

\item{bg.colour}{shadowtext background color}

\item{which.label}{which label to plot text}

\item{wc_alpha}{alpha value for wordcloud}

\item{wcScale}{scaling value for wordcloud}

\item{rot.per}{ggwordcloud parameter}

\item{rad}{named vector of size of each cluster}

\item{top}{Top-{top} genes are included}

\item{gene_name}{show gene names instead of textual information}

\item{withggfx}{applying ggfx filters}

\item{ggfxParams}{parameter list for ggfx}

\item{sortBy}{default to avg_log2FC, "log10p" can be specified.}

\item{scaleNumber}{scale the frequency of words by this number
in `gene_name`}

\item{decreasing}{sort by decreasing order or not}

\item{geneNum}{number of genes to be included in wordclouds}

\item{random.order}{ggwordcloud parameter}

\item{dimred}{dimension reduction method}

\item{base_ellipse}{if TRUE, wordclouds are placed based on \code{stat_ellipse}.}

\item{base_dens}{if TRUE, wordclouds are placed based on density}

\item{withTitle}{whether to append title on wordcloud}

\item{args}{parameters to passed to refseq}
}
\value{
single-cell plot with text annotation
}
\description{
plotReducedDimWithTexts
}
