% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotextgraph.R
\name{plotWC}
\alias{plotWC}
\alias{plotWC,biotext-method}
\title{plotWC}
\usage{
plotWC(
  x,
  tagPalette = NULL,
  madeUpper = c("dna", "rna"),
  preserve = FALSE,
  scaleFreq = NULL,
  fontFamily = "sans",
  numWords = NULL,
  wcScale = 10,
  argList = list(),
  useggwordcloud = TRUE,
  asis = FALSE
)

\S4method{plotWC}{biotext}(
  x,
  tagPalette = NULL,
  madeUpper = c("dna", "rna"),
  preserve = FALSE,
  scaleFreq = NULL,
  fontFamily = "sans",
  numWords = NULL,
  wcScale = 10,
  argList = list(),
  useggwordcloud = TRUE,
  asis = FALSE
)
}
\arguments{
\item{x}{biotext object}

\item{tagPalette}{tag palette when `tag` is TRUE. It is also used for dependency network
using udpipe, and tagging colorization for word cloud.
Default to NULL, which indicates automatically set.}

\item{madeUpper}{make these words uppercase in resulting plot,
default to c("rna" and "dna")}

\item{preserve}{Try to preserve original characters.}

\item{scaleFreq}{default to NULL, scale the value if specified}

\item{fontFamily}{font family to use, default to "sans".}

\item{numWords}{the number of words to be shown in the plot.
When `autoThresh` is TRUE, the number of this value will be shown.}

\item{wcScale}{scaling size for ggwordcloud}

\item{argList}{parameters to pass to wordcloud() or ggwordcloud()}

\item{useggwordcloud}{default to TRUE, otherwise use `wordcloud` function.}

\item{asis}{plot the original network (default to FALSE)}
}
\value{
wordcloud visualization
}
\description{
Plot the wordcloud changing the visualization parameters.
}
\details{
The function accepts the already calculated biotext object and outputs the visualization based on
the specified parameters.
}
\examples{
geneList <- c("DDX41","PNKP","ERCC1","IRF3","XRCC1")
test <- refseq(geneList, plotType="wc")
plotWC(test, asis=TRUE)
}
