% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{process_network_manual}
\alias{process_network_manual}
\title{process_network_manual}
\usage{
process_network_manual(
  ret,
  delete_zero_degree = TRUE,
  make_upper = NULL,
  query_plot = FALSE,
  drop_ID = TRUE
)
}
\arguments{
\item{ret}{biotext class object}

\item{delete_zero_degree}{delete zero degree nodes}

\item{make_upper}{make these words uppercase}

\item{query_plot}{plot the column other than text}

\item{drop_ID}{drop `ID` column}
}
\value{
biotext class object
}
\description{
process graph made from user-specified df
}
\examples{
obtain_manual(data.frame(text=c("test4 test4",
                                "test3 test3 testi",
                                "test2 test2"))) |>
make_corpus(num_only=TRUE) |>
make_TDM() |>
make_graph() |>
process_network_manual()

}
