% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refseq.R
\name{refseq}
\alias{refseq}
\title{refseq}
\usage{
refseq(
  geneList,
  keyType = "SYMBOL",
  excludeFreq = 2000,
  exclude = "frequency",
  filterMax = FALSE,
  excludeType = ">",
  tfidf = FALSE,
  genePlotNum = 10,
  preserve = FALSE,
  takeMax = FALSE,
  additionalRemove = NA,
  onlyCorpus = FALSE,
  madeUpper = c("dna", "rna"),
  organism = 9606,
  pal = c("blue", "red"),
  numWords = 30,
  scaleRange = c(5, 10),
  autoScale = FALSE,
  showLegend = FALSE,
  orgDb = org.Hs.eg.db,
  edgeLabel = FALSE,
  naEdgeColor = "grey50",
  cooccurrence = FALSE,
  pvclAlpha = 0.95,
  cl = FALSE,
  ngram = 1,
  plotType = "network",
  onlyTDM = FALSE,
  stem = FALSE,
  colorText = FALSE,
  corThresh = 0.2,
  genePlot = FALSE,
  autoThresh = TRUE,
  autoNumWords = FALSE,
  genePathPlot = NULL,
  genePathPlotSig = 0.05,
  tag = "none",
  layout = "nicely",
  edgeLink = TRUE,
  deleteZeroDeg = TRUE,
  enrich = NULL,
  topPath = 10,
  ora = FALSE,
  tagWhole = FALSE,
  mergeCorpus = NULL,
  numOnly = TRUE,
  madeUpperGenes = TRUE,
  onWholeDTM = FALSE,
  pre = TRUE,
  takeMean = FALSE,
  tagPalette = NULL,
  collapse = FALSE,
  addFreqToGene = FALSE,
  useUdpipe = FALSE,
  normalize = FALSE,
  fontFamily = "sans",
  udpipeModel = "english-ewt-ud-2.5-191206.udpipe",
  scaleFreq = NULL,
  colorize = FALSE,
  geneColor = "grey",
  argList = list(),
  useggwordcloud = TRUE,
  wcScale = 10,
  catColors = NULL,
  discreteColorWord = FALSE,
  useSeed = 42,
  scaleEdgeWidth = c(1, 3),
  splitByEA = NULL,
  filterByGO = FALSE,
  docsum = FALSE
)
}
\arguments{
\item{geneList}{gene ID list}

\item{keyType}{default to SYMBOL}

\item{excludeFreq}{default to 5000}

\item{exclude}{"frequency" or "tfidf",}

\item{filterMax}{Use pre-calculated filter based on max-values when excluding TfIdf
Otherwise take sum.}

\item{excludeType}{">" or "<", combined with `exclude` and `excludeFreq`,
e.g. filter the words with the pre-calculated frequency > 5000}

\item{tfidf}{use TfIdf when making TDM, default to FALSE.}

\item{genePlotNum}{number of genes to be plotted (default: 10)}

\item{preserve}{Try to preserve original characters.}

\item{takeMax}{Take max values for each term in term-document matrix}

\item{additionalRemove}{specific words to be excluded}

\item{onlyCorpus}{return only corpus (tm).}

\item{madeUpper}{make these words uppercase in resulting plot,
default to c("rna" and "dna")}

\item{organism}{organism ID to use in `GeneSummary`}

\item{pal}{palette for color gradient in correlation network.
should be a vector of length two like c("red","blue").}

\item{numWords}{the number of words to be shown in the plot.
When `autoThresh` is TRUE, the number of this value will be shown.}

\item{scaleRange}{scale for label and node size in the network.}

\item{autoScale}{scale the label and node size automatically for the large network.}

\item{showLegend}{whether to show legend in the network}

\item{orgDb}{the database used to convert identifiers, default to org.Hs.eg.db.}

\item{edgeLabel}{if TRUE, plot the edge label (default: FALSE)}

\item{naEdgeColor}{edge colors for NA values (linking query with the category other than text)}

\item{cooccurrence}{default to FALSE, if TRUE, use cooccurrence instead of correlation.}

\item{pvclAlpha}{alpha value for pvpick()}

\item{cl}{for parPvclust, parallel clustering can be performed}

\item{ngram}{N-gram specification, default to 1.}

\item{plotType}{"wc" or "network", default to "network"}

\item{onlyTDM}{return only TDM (tm).}

\item{stem}{whether to use stemming when making corpus.}

\item{colorText}{color text label based on frequency in the network}

\item{corThresh}{the correlation (cooccurrence) threshold.}

\item{genePlot}{plot associated genes (default: FALSE)}

\item{autoThresh}{automatically choose thresholding value to show the `numWords`,
when deleteZeroDeg (deleting no-connected words) is TRUE, which is default.}

\item{autoNumWords}{determine the number of words to be shown by ORA, default to FALSE.}

\item{genePathPlot}{plot associated genes and pathways (default: NULL)
Must be "kegg" or "reactome", automatically set genePlot to TRUE.}

\item{genePathPlotSig}{threshold for adjusted p-values (default: 0.05)}

\item{tag}{perform pvclust on words and colorlize them in wordcloud or network
argument of those accepted in pvclust `method.dist` option, like "correlation".
Default to "none", which performs no tagging.}

\item{layout}{the layout for the network, defaul to "nicely".
It can be one of the layouts implemented in `igraph` and `ggraph`, such as
`kk` (Kamada-Kawai), `nicely` (automatic selection of algorithm), `drl` (the force-directed DrL layout).
The options are available at: https://igraph.org/r/doc/layout_.html}

\item{edgeLink}{if FALSE, use geom_edge_diagonal. if TRUE, geom_edge_link. Default to TRUE.}

\item{deleteZeroDeg}{delete zero degree node from plot in the network}

\item{enrich}{currently, only 'reactome' and 'kegg' is supported.}

\item{topPath}{how many pathway descriptions are included in text analysis,
sorted by p-values in the results.}

\item{ora}{perform over-representation analysis or not (experimental)}

\item{tagWhole}{whether to perform pvclust on whole matrix or subset of the matrix.}

\item{mergeCorpus}{specify multiple corpus if intend to combine them.
like PubMed information and RefSeq summary}

\item{numOnly}{delete number only (not deleting XXX123, but delete only the number)}

\item{madeUpperGenes}{make genes upper case automatically (default to TRUE)
This uses the `SYMBOL` key in `orgDb`.}

\item{onWholeDTM}{calculate correlation network
on whole dataset or top-words specified by numWords.}

\item{pre}{remove preset filtering words.}

\item{takeMean}{take mean values for each term in term-document matrix.}

\item{tagPalette}{tag palette when `tag` is TRUE. It is also used for dependency network
using udpipe, and tagging colorization for word cloud.
Default to NULL, which indicates automatically set.}

\item{collapse}{default to FALSE, collapse all the sentences.}

\item{addFreqToGene}{add pseudo frequency to gene in genePlot, default to FALSE.}

\item{useUdpipe}{use udpipe to make a dependency network.}

\item{normalize}{sum normalize the term frequency document-wise.}

\item{fontFamily}{font family to use, default to "sans".}

\item{udpipeModel}{udpipe model file name.}

\item{scaleFreq}{default to NULL, scale the value if specified}

\item{colorize}{color the word nodes by their frequency, and the other nodes by their category.
if colorize=FALSE and addFreqToGene=TRUE, gene nodes are colorized according to the minimum frequency 
of the words in the network}

\item{geneColor}{color for associated genes with words (used when tag or colorize option is TRUE)}

\item{argList}{parameters to pass to wordcloud() or ggwordcloud()}

\item{useggwordcloud}{default to TRUE, otherwise use `wordcloud` function.}

\item{wcScale}{scaling size for ggwordcloud}

\item{catColors}{colors for words and texts when colorize is TRUE and discreteColorWord is TRUE}

\item{discreteColorWord}{colorize words by "Words" category, not frequency.}

\item{useSeed}{random seed}

\item{scaleEdgeWidth}{scale for edge width}

\item{splitByEA}{automatically split the genes based on significant enrichment analysis results,
and returns the list of object for each term. Default to NULL. Must be 'kegg' or 'reactome',
in which the function performs over-representation analysis by enrichKEGG or enrichPathway in
clusterProfiler and ReactomePA.}

\item{filterByGO}{filter the results to the words obtained from GO terms,
while preserving the number of words to be shown}

\item{docsum}{if TRUE, convert the term-document matrix to binary.}
}
\value{
`biotext` class object
}
\description{
Text mining RefSeq description obtained by GeneSummary
}
\examples{
geneList <- c("DDX41","PNKP")
refseq(geneList)

}
