% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWordsOnDendro.R
\name{returnPyramid}
\alias{returnPyramid}
\title{returnPyramid}
\usage{
returnPyramid(
  L,
  R,
  geneVec,
  geneVecType,
  filterWords,
  numberOfWords,
  orgDb,
  additionalRemove = NA,
  widths = c(0.3, 0.3, 0.3),
  lowCol = "blue",
  tfidf = TRUE,
  highCol = "red",
  madeUpper = c("rna", "dna")
)
}
\arguments{
\item{L}{genes in the cluster}

\item{R}{genes in the other cluster}

\item{geneVec}{gene-named vector of node names in dendrogram}

\item{geneVecType}{type of the name of geneVec (default: ENSEMBL)}

\item{filterWords}{words to filter based on frequency}

\item{numberOfWords}{the number of words to plot}

\item{orgDb}{database to change the ID to ENTREZ ID}

\item{additionalRemove}{words to filter}

\item{widths}{parameters to pass to patchwork}

\item{lowCol}{gradient low color}

\item{tfidf}{use tfidf}

\item{highCol}{gradient high color}

\item{madeUpper}{words with uppercase}
}
\value{
list of pyramid plot grobs and its positions
}
\description{
Return pyramid plots
}
