% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWordsOnDendro.R
\name{returnPyramid}
\alias{returnPyramid}
\title{returnPyramid}
\usage{
returnPyramid(
  L,
  R,
  geneVec,
  geneVecType,
  numberOfWords = 25,
  widths = c(0.3, 0.3, 0.3),
  lowCol = "blue",
  showType = "ID",
  highCol = "red",
  highlight = NULL,
  type = "words",
  wrap = 15,
  orgDb = org.Hs.eg.db,
  argList = list()
)
}
\arguments{
\item{L}{genes in the cluster}

\item{R}{genes in the other cluster}

\item{geneVec}{gene-named vector of node names in dendrogram}

\item{geneVecType}{type of the name of geneVec (default: ENSEMBL)}

\item{numberOfWords}{the number of words to plot}

\item{widths}{parameters to pass to patchwork}

\item{lowCol}{gradient low color}

\item{showType}{which labels to show in enrich}

\item{highCol}{gradient high color}

\item{highlight}{words to highlight}

\item{type}{"words" or "enrich"}

\item{orgDb}{organism database to use in enrich}

\item{argList}{parameters passed to wcGeneSummary()}
}
\value{
list of pyramid plot grobs and its positions
}
\description{
Return pyramid plots
}
