% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWordsOnDendro.R
\name{returnPyramid}
\alias{returnPyramid}
\title{returnPyramid}
\usage{
returnPyramid(
  L,
  R,
  geneVec,
  geneVecType,
  numberOfWords = 25,
  widths = c(0.3, 0.6, 0.3),
  lowCol = "blue",
  showType = "ID",
  useRandomColor = FALSE,
  highCol = "red",
  highlight = NULL,
  wcScale = 3,
  type = "words",
  wrap = 15,
  textSize = 3.5,
  useggfx = NULL,
  takeIntersect = TRUE,
  useWC = FALSE,
  wcArgs = list(),
  bg.colour = NULL,
  orgDb = org.Hs.eg.db,
  argList = list(),
  useFunc = NULL,
  useDf = NULL
)
}
\arguments{
\item{L}{genes in the cluster}

\item{R}{genes in the other cluster}

\item{geneVec}{gene-named vector of node names in dendrogram}

\item{geneVecType}{type of the name of geneVec (default: ENSEMBL)}

\item{numberOfWords}{the number of words to plot}

\item{widths}{parameters to pass to patchwork}

\item{lowCol}{gradient low color}

\item{showType}{which labels to show in enrich}

\item{useRandomColor}{use random colors on wordclouds}

\item{highCol}{gradient high color}

\item{highlight}{words to highlight}

\item{wcScale}{if useWC, number of size scaling (max_size)}

\item{type}{"words" or "enrich"}

\item{wrap}{wrap the strings}

\item{textSize}{text size in pyramid plots}

\item{useggfx}{use ggfx on plot}

\item{takeIntersect}{take intersection or frequent words}

\item{useWC}{return wordcloud}

\item{wcArgs}{argument list for ggwordcloud}

\item{bg.colour}{background color for wordclouds}

\item{orgDb}{organism database to use in enrich}

\item{argList}{parameters passed to wcGeneSummary()}

\item{useFunc}{function to summarize text}

\item{useDf}{data.frame to subset when manual function is specified}
}
\value{
list of pyramid plot grobs and its positions
}
\description{
Return pyramid plots
}
