% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{returnSim}
\alias{returnSim}
\title{returnSim}
\usage{
returnSim(cllist, keyType = "ENTREZID", numLimit = 5000, argList = list())
}
\arguments{
\item{cllist}{cluster list (named vector or list)}

\item{keyType}{keytype}

\item{numLimit}{threshold for gene number limit
default to 5000}

\item{argList}{parameters to pass to wcGeneSummary()}
}
\value{
similarity matrix
}
\description{
return similarity matrix of cluster
based on intersection over union of words
}
\examples{
ex <- returnExample()
returnSim(ex$color, keyType="ENSEMBL")
}
