% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.R
\name{TextMarkersScran}
\alias{TextMarkersScran}
\title{TextMarkersScran}
\usage{
TextMarkersScran(
  res,
  keyType = "SYMBOL",
  type = "wc",
  genePlot = TRUE,
  genePlotNum = 5,
  colorText = TRUE,
  args = list(),
  wcArgs = NULL,
  raw = FALSE,
  col = NULL,
  withTitle = TRUE,
  top = 5,
  withggfx = NULL,
  FDRThresh = 0.05,
  ggfxParams = list()
)
}
\arguments{
\item{res}{result of findMarkers()}

\item{keyType}{keytype}

\item{type}{wc or network}

\item{genePlot}{whether to plot relevant genes}

\item{genePlotNum}{number of genes to plot}

\item{colorText}{colorlize text or not, default to TRUE}

\item{args}{parameters to pass to refseq}

\item{wcArgs}{parameters to pass to ggwordcloud}

\item{raw}{obtain raw results of refseq instead of plot}

\item{col}{color to be used in wordcloud}

\item{withTitle}{ggtitle({cluster name}) will be added}

\item{top}{Top-{top} genes for each cluster will be included}

\item{withggfx}{applying ggfx filters}

\item{FDRThresh}{FDR threshold}

\item{ggfxParams}{parameter list for ggfx}
}
\value{
list of ggplot
}
\description{
Take results of scran::findMarkers as input,
and plot wordcloud or network for all the clusters.
}
\examples{
df <- data.frame(
  p.value=c(0.01, 0.01),gene=c("PNKP","DDX41"),
  Top=c(1,2),FDR=c(0.01, 0.01)
)
row.names(df) <- df$gene
markers <- list("1"=df)
colors <- list("1"="blue")
TextMarkersScran(markers, col=colors)

}
