% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textORA.R
\name{textORA}
\alias{textORA}
\title{textORA}
\usage{
textORA(queries)
}
\arguments{
\item{queries}{gene list (Entrez ID)}
}
\value{
p-values for the words

p-values for the words
}
\description{
Return p-values from hypergeometric distribution
}
\examples{
textORA(c("2067","2068","2071","2072"))
}
