% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcAbst.R
\name{wcAbst}
\alias{wcAbst}
\title{wcAbst}
\usage{
wcAbst(
  queries,
  redo = NULL,
  madeUpper = c("dna", "rna"),
  target = "abstract",
  useFil = NA,
  filType = "above",
  filNum = 0,
  sortOrder = "relevance",
  pvclAlpha = 0.95,
  numOnly = TRUE,
  delim = "OR",
  limit = 10,
  geneUpper = TRUE,
  apiKey = NULL,
  tfidf = FALSE,
  cl = FALSE,
  pal = c("blue", "red"),
  numWords = 30,
  scaleRange = c(5, 10),
  showLegend = FALSE,
  plotType = "wc",
  colorText = FALSE,
  quote = FALSE,
  corThresh = 0.2,
  layout = "nicely",
  tag = FALSE,
  tagWhole = FALSE,
  onlyCorpus = FALSE,
  onlyTDM = FALSE,
  bn = FALSE,
  R = 20,
  retMax = 10,
  edgeLabel = FALSE,
  edgeLink = TRUE,
  ngram = NA,
  genePlot = FALSE,
  onlyDf = FALSE,
  nodePal = palette(),
  preserve = TRUE,
  takeMax = FALSE,
  useUdpipe = FALSE,
  udpipeOnlyFreq = FALSE,
  udpipeOnlyFreqN = FALSE,
  udpipeModel = "english-ewt-ud-2.5-191206.udpipe",
  normalize = FALSE,
  takeMean = FALSE,
  deleteZeroDeg = TRUE,
  additionalRemove = NA,
  orgDb = org.Hs.eg.db,
  onlyGene = FALSE,
  pre = FALSE,
  onWholeDTM = FALSE,
  madeUpperGenes = TRUE,
  stem = FALSE,
  argList = list()
)
}
\arguments{
\item{queries}{gene symbols}

\item{redo}{if plot in other parameters, input the previous list}

\item{madeUpper}{make the words uppercase in resulting plot}

\item{target}{"abstract" or "title"}

\item{useFil}{filter based on "GS_TfIdf" (whole gene summary tf-idf)
or "BSDB_TfIdf" (whole bugsigdb tf-idf)}

\item{filType}{"above" or "below"}

\item{filNum}{specify filter tfidf}

\item{sortOrder}{sort order, passed to rentrez function}

\item{pvclAlpha}{alpha for pvpick()}

\item{numOnly}{delete number only}

\item{delim}{delimiter for queries}

\item{limit}{limit number for query count, default to 10}

\item{geneUpper}{make queries uppercase}

\item{apiKey}{api key for eutilities}

\item{tfidf}{use TfIdf when making TDM}

\item{cl}{cluster to pass to pvclust (snow::makeCluster(n))}

\item{pal}{palette for color gradient in correlation network}

\item{numWords}{the number of words to be shown}

\item{scaleRange}{scale for label and node size in correlation network}

\item{showLegend}{whether to show legend in correlation network}

\item{plotType}{"wc" or "network"}

\item{colorText}{color text label based on frequency in correlation network}

\item{quote}{whether to quote the queries}

\item{corThresh}{the correlation threshold}

\item{layout}{the layout for correlation network, defaul to "nicely"}

\item{tag}{cluster the words based on text using pvclust}

\item{tagWhole}{tag based on whole data or subset}

\item{onlyCorpus}{return only corpus}

\item{onlyTDM}{return only TDM}

\item{bn}{perform bootstrap-based Bayesian network inference 
instead of correlation using bnlearn}

\item{R}{how many bootstrap when bn is stated}

\item{retMax}{how many items are to be retlieved?}

\item{edgeLabel}{if TRUE, plot the edge label (default: FALSE)}

\item{edgeLink}{if FALSE, use geom_edge_diagonal}

\item{ngram}{default to NA (1)}

\item{genePlot}{plot associated genes (default: FALSE)
Query gene name is shown with (Q)}

\item{onlyDf}{return only the raw data.frame of searching PubMed}

\item{nodePal}{node palette when tag is TRUE}

\item{preserve}{try to preserve the original characters}

\item{takeMax}{when summarizing term-document matrix, take max.
Otherwise take sum.}

\item{useUdpipe}{use udpipe to make a network}

\item{udpipeOnlyFreq}{when using udpipe, include only high-frequent words}

\item{udpipeOnlyFreqN}{when using udpipe, include only the neighbors of
high-frequent words}

\item{udpipeModel}{udpipe model file name}

\item{normalize}{sum normalize the term frequency document-wise}

\item{takeMean}{take mean values for each term in term-document matrix}

\item{deleteZeroDeg}{delete zero degree node from plot in correlation network}

\item{additionalRemove}{specific words to be excluded}

\item{orgDb}{org database, default to org.Hs.eg.db}

\item{onlyGene}{plot only the gene symbol
(orgDb with SYMBOL key can be used)}

\item{pre}{filter preset words like publisher's names}

\item{onWholeDTM}{calculate correlation network
on whole dataset or top-words specified by numWords}

\item{madeUpperGenes}{make genes upper case automatically (default to TRUE)}

\item{stem}{whether to use stemming}

\item{argList}{parameters to pass to wordcloud()}
}
\value{
list of data frame and ggplot2 object
}
\description{
make word cloud or correlation network from PubMed
}
\examples{
\dontrun{wcAbst("DDX41")}
}
