% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcAbst.R
\name{wcAbst}
\alias{wcAbst}
\title{wcAbst}
\usage{
wcAbst(
  queries,
  redo = NA,
  madeUpper = c("dna", "rna"),
  target = "abstract",
  usefil = NA,
  filnum = 0,
  pvclAlpha = 0.95,
  numOnly = TRUE,
  delim = "OR",
  geneUpper = TRUE,
  apiKey = NULL,
  tfidf = FALSE,
  pal = c("blue", "red"),
  numWords = 30,
  scaleRange = c(5, 10),
  showLegend = FALSE,
  plotType = "wc",
  colorText = FALSE,
  corThresh = 0.2,
  layout = "nicely",
  tag = FALSE,
  tagWhole = FALSE,
  onlyCorpus = FALSE,
  onlyTDM = FALSE,
  bn = FALSE,
  R = 20,
  retMax = 10,
  edgeLabel = FALSE,
  edgeLink = TRUE,
  ngram = NA,
  genePlot = FALSE,
  deleteZeroDeg = TRUE,
  additionalRemove = NA,
  orgDb = org.Hs.eg.db,
  preset = FALSE,
  onWholeDTM = FALSE,
  madeUpperGenes = TRUE,
  ...
)
}
\arguments{
\item{queries}{gene symbols (max: 5)}

\item{redo}{if plot in other parameters, input the previous list}

\item{madeUpper}{make the words uppercase in resulting plot}

\item{target}{"abstract" or "title"}

\item{usefil}{filter based on "gstfidf" (whole gene summary tf-idf)
or "bsdbtfidf" (whole bugsigdb tf-idf)}

\item{filnum}{specify filter tfidf}

\item{pvclAlpha}{alpha for pvpick()}

\item{numOnly}{delete number only}

\item{delim}{delimiter for queries}

\item{geneUpper}{make queries uppercase}

\item{apiKey}{api key for eutilities}

\item{tfidf}{use TfIdf when making TDM}

\item{pal}{palette for color gradient in correlation network}

\item{numWords}{the number of words to be shown}

\item{scaleRange}{scale for label and node size in correlation network}

\item{showLegend}{whether to show legend in correlation network}

\item{plotType}{"wc" or "network"}

\item{colorText}{color text label based on frequency in correlation network}

\item{corThresh}{the correlation threshold}

\item{layout}{the layout for correlation network, defaul to "nicely"}

\item{tag}{cluster the words based on text using pvclust}

\item{tagWhole}{tag based on whole data or subset}

\item{onlyCorpus}{return only corpus}

\item{onlyTDM}{return only TDM}

\item{bn}{perform bootstrap-based Bayesian network inference 
instead of correlation using bnlearn}

\item{R}{how many bootstrap when bn is stated}

\item{retMax}{how many items are to be retlieved?}

\item{edgeLabel}{if TRUE, plot the edge label (default: FALSE)}

\item{edgeLink}{if FALSE, use geom_edge_diagonal}

\item{ngram}{default to NA (1)}

\item{genePlot}{plot associated genes (default: FALSE)}

\item{deleteZeroDeg}{delete zero degree node from plot in correlation network}

\item{additionalRemove}{specific words to be excluded}

\item{orgDb}{org database, default to org.Hs.eg.db}

\item{preset}{filter preset words}

\item{onWholeDTM}{calculate correlation network
on whole dataset or top-words specified by numWords}

\item{madeUpperGenes}{make genes upper case automatically (default to TRUE)}

\item{...}{parameters to pass to wordcloud()}
}
\value{
list of data frame and ggplot2 object
}
\description{
make word cloud or correlation network from PubMed abstracts using easyPubMed
}
\examples{
wcAbst("DDX41")
}
