% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcBSDB.R
\name{wcBSDB}
\alias{wcBSDB}
\title{wcBSDB}
\usage{
wcBSDB(
  mbList,
  excludeFreq = 1000,
  exclude = "frequency",
  excludeType = ">",
  additionalRemove = NA,
  tfidf = FALSE,
  target = "title",
  apiKey = NULL,
  takeMax = FALSE,
  pre = FALSE,
  pvclAlpha = 0.95,
  numOnly = TRUE,
  madeUpper = c("dna", "rna"),
  redo = NULL,
  pal = c("blue", "red"),
  numWords = 15,
  preserve = TRUE,
  metab = NULL,
  metabThresh = 0.2,
  curate = TRUE,
  abstArg = list(),
  nodePal = palette(),
  metCol = NULL,
  scaleRange = c(5, 10),
  showLegend = FALSE,
  ecPlot = FALSE,
  edgeLabel = FALSE,
  mbPlot = FALSE,
  onlyTDM = FALSE,
  ecFile = NULL,
  upTaxFile = NULL,
  filterMax = FALSE,
  ngram = NA,
  plotType = "wc",
  disPlot = FALSE,
  onWholeDTM = FALSE,
  colorText = FALSE,
  corThresh = 0.2,
  tag = FALSE,
  tagWhole = FALSE,
  stem = FALSE,
  layout = "nicely",
  edgeLink = TRUE,
  deleteZeroDeg = TRUE,
  cl = FALSE,
  ...
)
}
\arguments{
\item{mbList}{microbe list}

\item{excludeFreq}{default to 1000 (no words are excluded)}

\item{exclude}{"frequency" or "tfidf",}

\item{excludeType}{">" or "<"}

\item{additionalRemove}{specific words to be excluded}

\item{tfidf}{use TfIdf when making TDM}

\item{target}{"title" or "abstract"}

\item{apiKey}{api key for eutilities}

\item{takeMax}{when summarizing term-document matrix, take max.}

\item{pre}{predefined filter words}

\item{pvclAlpha}{alpha for pvpick()}

\item{numOnly}{delete number only}

\item{madeUpper}{make the words uppercase in resulting plot}

\item{redo}{if "abstract" is chosen in target, one can provide resulting object again}

\item{pal}{palette for color gradient in correlation network and tag coloring}

\item{numWords}{the number of words to be shown}

\item{preserve}{try to preserve the original characters}

\item{metab}{tibble of metabolite - taxon association}

\item{metabThresh}{threshold of association}

\item{curate}{include articles in bugsigdb}

\item{abstArg}{passed to PubMed function when using curate=FALSE}

\item{nodePal}{node palette when tag is TRUE}

\item{metCol}{metabolite data frame column name in the order of
"candidate taxon", "metabolite", "quantitative values for thresholding"}

\item{scaleRange}{scale for label and node size in correlation network}

\item{showLegend}{whether to show legend in correlation network}

\item{ecPlot}{plot link between enzyme and microbes
this option requires two files to be passed to wcEC() and getUPTax().}

\item{edgeLabel}{if TRUE, plot the edge label (default: FALSE)}

\item{mbPlot}{plot microbe names}

\item{onlyTDM}{return only TDM}

\item{ecFile}{enzyme database file}

\item{upTaxFile}{UniProt taxonomy file}

\item{filterMax}{use pre-calculated filter based on max-values when excluding TfIdf
Otherwise take sum.}

\item{ngram}{default to NA (1)}

\item{plotType}{"wc" or "network"}

\item{disPlot}{plot diseases}

\item{onWholeDTM}{calculate correlation network
on whole dataset or top-words specified by numWords}

\item{colorText}{color text label based on frequency in correlation network}

\item{corThresh}{the correlation threshold}

\item{tag}{perform pvclust on words and colorlize them in wordcloud}

\item{tagWhole}{tag whole set or subset}

\item{stem}{whether to use stemming}

\item{layout}{the layout for correlation network, defaul to "nicely"}

\item{edgeLink}{if FALSE, use geom_edge_diagonal}

\item{deleteZeroDeg}{delete zero degree node from plot in correlation network}

\item{cl}{cluster object passed to pvclust}

\item{...}{parameters to pass to wordcloud()}
}
\value{
list of data frame and ggplot2 object
}
\description{
Visualize BugSigDB
}
\examples{
mbList <- c("Veillonella dispar")
wcBSDB(mbList)
}
