% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcEC.R
\name{wcEC}
\alias{wcEC}
\title{wcEC}
\usage{
wcEC(
  file,
  ecnum,
  onlyTerm = FALSE,
  onlyDf = FALSE,
  taxec = FALSE,
  taxFile = NULL,
  candTax = NULL,
  argList = list()
)
}
\arguments{
\item{file}{file downloaded from expasy}

\item{ecnum}{candidate ecnum, like those obtained from eggNOG-mapper}

\item{onlyTerm}{only return quoted queries to wcAbst}

\item{onlyDf}{only return ec description data.frame
if onlyTerm and onlyDf are both specified, onlyTerm have priority}

\item{taxec}{link taxonomy to EC using UniProt Taxonomy ID file
If this is TRUE, data.frame is returned}

\item{taxFile}{UniProt organism ID file path}

\item{candTax}{when taxec=TRUE, search only for these species.}

\item{argList}{passed to osplot(target="pubmed")}
}
\value{
object consisting of data frame and ggplot2 object
}
\description{
Query the Enzyme Comission number and obtain description,
and search pubmed for these enzymes and make word cloud and
correlation network. Need "enzyme.dat" from ExPASy
(https://enzyme.expasy.org/).
}
\examples{
file <- "enzyme.dat"
\donttest{wcEC(file, ecnum="1.2.1.1")}
}
