% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcGeneSummary.R
\name{wcGeneSummary}
\alias{wcGeneSummary}
\title{wcGeneSummary}
\usage{
wcGeneSummary(
  geneList,
  excludeFreq = 5000,
  additionalRemove = NA,
  madeUpper = c("dna", "rna"),
  organism = 9606,
  palette = c("blue", "red"),
  numWords = 15,
  scaleRange = c(5, 10),
  ORA = FALSE,
  showLegend = FALSE,
  plotType = "wc",
  corThresh = 0.6,
  layout = "nicely",
  edgeLink = TRUE,
  deleteZeroDeg = TRUE,
  ...
)
}
\arguments{
\item{geneList}{ENTREZID list}

\item{excludeFreq}{exclude words with overall frequency above excludeFreq, default to 5000}

\item{additionalRemove}{specific words to be excluded}

\item{madeUpper}{make the words uppercase in resulting plot}

\item{palette}{palette for color gradient in correlation network}

\item{numWords}{the number of words to be shown in correlation network}

\item{scaleRange}{scale for label and node size in correlation network}

\item{ORA}{perform over-representation analysis}

\item{showLegend}{whether to show legend in correlation network}

\item{plotType}{"wc" or "network"}

\item{corThresh}{the correlation threshold}

\item{layout}{the layout for correlation network, defaul to "nicely"}

\item{edgeLink}{if FALSE, use geom_edge_diagonal}

\item{deleteZeroDeg}{delete zero degree node from plot in correlation network}
}
\value{
list of data frame and ggplot2 object
}
\description{
Plot wordcloud of RefSeq description obtained by GeneSummary
}
\examples{
wcGeneSummary(geneList)
}
