% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcGeneSummary.R
\name{wcGeneSummary}
\alias{wcGeneSummary}
\title{wcGeneSummary}
\usage{
wcGeneSummary(
  geneList,
  keyType = "SYMBOL",
  excludeFreq = 2000,
  excludeTfIdf = NA,
  tfidf = FALSE,
  genePlotNum = 10,
  additionalRemove = NA,
  onlyCorpus = FALSE,
  madeUpper = c("dna", "rna"),
  organism = 9606,
  pal = c("blue", "red"),
  numWords = 15,
  scaleRange = c(5, 10),
  showLegend = FALSE,
  orgDb = org.Hs.eg.db,
  edgeLabel = FALSE,
  pvclAlpha = 0.95,
  bn = FALSE,
  R = 20,
  cl = FALSE,
  ngram = NA,
  plotType = "wc",
  onlyTDM = FALSE,
  colorText = FALSE,
  corThresh = 0.2,
  genePlot = FALSE,
  genePathPlot = NA,
  genePathPlotSig = 0.05,
  tag = FALSE,
  layout = "nicely",
  edgeLink = TRUE,
  deleteZeroDeg = TRUE,
  enrich = NULL,
  topPath = 10,
  ora = FALSE,
  tagWhole = FALSE,
  mergeCorpus = NULL,
  numOnly = TRUE,
  madeUpperGenes = TRUE,
  onWholeDTM = FALSE,
  autoFilter = FALSE,
  verbPOS = c("VBZ"),
  verb = FALSE,
  udpipeModel = "english-ewt-ud-2.5-191206.udpipe",
  ...
)
}
\arguments{
\item{geneList}{gene ID list}

\item{keyType}{default to SYMBOL}

\item{excludeFreq}{exclude words with overall frequency above excludeFreq
default to 5000}

\item{excludeTfIdf}{exclude based on tfidf (default: NA)}

\item{tfidf}{use TfIdf when making TDM}

\item{genePlotNum}{number of genes to be plotted}

\item{additionalRemove}{specific words to be excluded}

\item{onlyCorpus}{return only corpus}

\item{madeUpper}{make the words uppercase in resulting plot}

\item{organism}{organism ID to use}

\item{pal}{palette for color gradient in correlation network and tag coloring}

\item{numWords}{the number of words to be shown}

\item{scaleRange}{scale for label and node size in correlation network}

\item{showLegend}{whether to show legend in correlation network}

\item{orgDb}{default to org.Hs.eg.db}

\item{edgeLabel}{if TRUE, plot the edge label (default: FALSE)}

\item{pvclAlpha}{alpha for pvpick()}

\item{bn}{perform bootstrap-based Bayesian network inference instead of correlation using bnlearn}

\item{R}{how many bootstrap when bn is stated}

\item{cl}{for parPvclust, parallel clustering can be performed}

\item{ngram}{default to NA (1)}

\item{plotType}{"wc" or "network"}

\item{onlyTDM}{return only TDF}

\item{colorText}{color text label based on frequency in correlation network}

\item{corThresh}{the correlation threshold}

\item{genePlot}{plot associated genes (default: FALSE)}

\item{genePathPlot}{plot associated genes and pathways (default: FALSE)
"kegg" or "reactome"}

\item{genePathPlotSig}{threshold for adjusted p-values (default: 0.05)}

\item{tag}{perform pvclust on words and colorlize them in wordcloud}

\item{layout}{the layout for correlation network, defaul to "nicely"}

\item{edgeLink}{if FALSE, use geom_edge_diagonal}

\item{deleteZeroDeg}{delete zero degree node from plot in correlation network}

\item{enrich}{currently, only 'reactome' and 'kegg' is supported}

\item{topPath}{how many pathway descriptions are included in text analysis}

\item{ora}{perform ora or not (experimental)}

\item{tagWhole}{whether to perform pvclust on whole matrix or subset}

\item{mergeCorpus}{specify multiple corpus if intend to combine them.
like PubMed information and RefSeq summary}

\item{numOnly}{delete number only (not deleting XXX123)}

\item{madeUpperGenes}{make genes upper case automatically (default to TRUE)}

\item{onWholeDTM}{calculate correlation network
on whole dataset or top-words specified by numWords}

\item{autoFilter}{remove words "pmids", "geneids"}

\item{verbPOS}{Part of Speech tags in verbs, like "VBN", "VBZ"}

\item{verb}{show verb on edges (using udpipe)
peform udpipe_download_model(language = "english") beforehand,
and path udpipe file to udpipeModel}

\item{udpipeModel}{udpipe model file name}

\item{...}{parameters to pass to wordcloud()}
}
\value{
list of data frame and ggplot2 object
}
\description{
Plot wordcloud of RefSeq description obtained by GeneSummary
}
\examples{
geneList <- c("DDX41")
wcGeneSummary(geneList)
}
