% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcGeneSummary.R
\name{wcGeneSummary}
\alias{wcGeneSummary}
\title{wcGeneSummary}
\usage{
wcGeneSummary(
  geneList,
  keyType = "SYMBOL",
  excludeFreq = 2000,
  exclude = "frequency",
  filterMax = FALSE,
  excludeType = ">",
  tfidf = FALSE,
  genePlotNum = 10,
  preserve = TRUE,
  takeMax = FALSE,
  additionalRemove = NA,
  onlyCorpus = FALSE,
  madeUpper = c("dna", "rna"),
  organism = 9606,
  pal = c("blue", "red"),
  numWords = 15,
  scaleRange = c(5, 10),
  showLegend = FALSE,
  orgDb = org.Hs.eg.db,
  edgeLabel = FALSE,
  naEdgeColor = "grey50",
  cooccurrence = FALSE,
  pvclAlpha = 0.95,
  bn = FALSE,
  R = 20,
  cl = FALSE,
  ngram = NA,
  plotType = "wc",
  onlyTDM = FALSE,
  stem = FALSE,
  colorText = FALSE,
  corThresh = 0.2,
  genePlot = FALSE,
  genePathPlot = NA,
  genePathPlotSig = 0.05,
  tag = FALSE,
  layout = "nicely",
  edgeLink = TRUE,
  deleteZeroDeg = TRUE,
  enrich = NULL,
  topPath = 10,
  ora = FALSE,
  tagWhole = FALSE,
  mergeCorpus = NULL,
  numOnly = TRUE,
  madeUpperGenes = TRUE,
  onWholeTDM = FALSE,
  pre = TRUE,
  takeMean = FALSE,
  nodePal = palette(),
  collapse = FALSE,
  addFreqToGene = FALSE,
  useUdpipe = FALSE,
  normalize = FALSE,
  fontFamily = "sans",
  udpipeModel = "english-ewt-ud-2.5-191206.udpipe",
  scaleLowFreq = NULL,
  colorize = FALSE,
  geneColor = "grey",
  argList = list(),
  useggwordcloud = TRUE,
  wcScale = 10
)
}
\arguments{
\item{geneList}{gene ID list}

\item{keyType}{default to SYMBOL}

\item{excludeFreq}{default to 5000}

\item{exclude}{"frequency" or "tfidf",}

\item{filterMax}{use pre-calculated filter based on max-values when excluding TfIdf
Otherwise take sum.}

\item{excludeType}{">" or "<"}

\item{tfidf}{use TfIdf when making TDM}

\item{genePlotNum}{number of genes to be plotted}

\item{preserve}{preserve original characters}

\item{takeMax}{take max values for each term in term-document matrix}

\item{additionalRemove}{specific words to be excluded}

\item{onlyCorpus}{return only corpus}

\item{madeUpper}{make the words uppercase in resulting plot}

\item{organism}{organism ID to use}

\item{pal}{palette for color gradient in correlation network and tag coloring}

\item{numWords}{the number of words to be shown}

\item{scaleRange}{scale for label and node size in correlation network}

\item{showLegend}{whether to show legend in correlation network}

\item{orgDb}{default to org.Hs.eg.db}

\item{edgeLabel}{if TRUE, plot the edge label (default: FALSE)}

\item{naEdgeColor}{edge colors for NA values (linking query with the category other than text)}

\item{cooccurrence}{default to FALSE, if TRUE, use cooccurrence instead of correlation}

\item{pvclAlpha}{alpha for pvpick()}

\item{bn}{perform bootstrap-based Bayesian network inference instead of correlation using bnlearn}

\item{R}{how many bootstrap when bn is stated}

\item{cl}{for parPvclust, parallel clustering can be performed}

\item{ngram}{default to NA (1)}

\item{plotType}{"wc" or "network"}

\item{onlyTDM}{return only TDF}

\item{stem}{whether to use stemming}

\item{colorText}{color text label based on frequency in correlation network}

\item{corThresh}{the correlation (cooccurrence) threshold}

\item{genePlot}{plot associated genes (default: FALSE)}

\item{genePathPlot}{plot associated genes and pathways (default: FALSE)
"kegg" or "reactome"}

\item{genePathPlotSig}{threshold for adjusted p-values (default: 0.05)}

\item{tag}{perform pvclust on words and colorlize them in wordcloud}

\item{layout}{the layout for correlation network, defaul to "nicely"}

\item{edgeLink}{if FALSE, use geom_edge_diagonal}

\item{deleteZeroDeg}{delete zero degree node from plot in correlation network}

\item{enrich}{currently, only 'reactome' and 'kegg' is supported}

\item{topPath}{how many pathway descriptions are included in text analysis}

\item{ora}{perform ora or not (experimental)}

\item{tagWhole}{whether to perform pvclust on whole matrix or subset}

\item{mergeCorpus}{specify multiple corpus if intend to combine them.
like PubMed information and RefSeq summary}

\item{numOnly}{delete number only (not deleting XXX123)}

\item{madeUpperGenes}{make genes upper case automatically (default to TRUE)}

\item{onWholeTDM}{calculate correlation network
on whole dataset or top-words specified by numWords}

\item{pre}{remove words "pmids", "geneids"}

\item{takeMean}{take mean values for each term in term-document matrix}

\item{nodePal}{node palette when tag is TRUE}

\item{collapse}{default to FALSE, collapse all the sentences}

\item{addFreqToGene}{add pseudo frequency to gene in genePlot}

\item{useUdpipe}{use udpipe to make a network}

\item{normalize}{sum normalize the term frequency document-wise}

\item{fontFamily}{font family to use, default to "sans"}

\item{udpipeModel}{udpipe model file name}

\item{scaleLowFreq}{default to NULL, scale the value if specified}

\item{colorize}{color the nodes and texts based on their category}

\item{geneColor}{color for associated genes with words}

\item{argList}{parameters to pass to wordcloud()}

\item{useggwordcloud}{default to TRUE}

\item{wcScale}{scaling size for ggwordcloud}
}
\value{
list of data frame and ggplot2 object
}
\description{
Plot wordcloud of RefSeq description obtained by GeneSummary
Testing function for new class
}
\examples{
geneList <- c("DDX41")
wcGeneSummary(geneList)

}
