% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcGeneSummary.R
\name{wcGeneSummary}
\alias{wcGeneSummary}
\title{wcGeneSummary}
\usage{
wcGeneSummary(
  geneList,
  keyType = "SYMBOL",
  excludeFreq = 5000,
  additionalRemove = NA,
  madeUpper = c("dna", "rna"),
  organism = 9606,
  palette = c("blue", "red"),
  numWords = 15,
  scaleRange = c(5, 10),
  showLegend = FALSE,
  orgDb = org.Hs.eg.db,
  plotType = "wc",
  colorText = FALSE,
  corThresh = 0.6,
  layout = "nicely",
  edgeLink = TRUE,
  deleteZeroDeg = TRUE,
  enrich = NULL,
  topPath = 10,
  ora = FALSE,
  ...
)
}
\arguments{
\item{geneList}{ENTREZID list}

\item{keyType}{default to SYMBOL}

\item{excludeFreq}{exclude words with overall frequency above excludeFreq
default to 5000}

\item{additionalRemove}{specific words to be excluded}

\item{madeUpper}{make the words uppercase in resulting plot}

\item{organism}{organism ID to use}

\item{palette}{palette for color gradient in correlation network}

\item{numWords}{the number of words to be shown}

\item{scaleRange}{scale for label and node size in correlation network}

\item{showLegend}{whether to show legend in correlation network}

\item{orgDb}{default to org.Hs.eg.db}

\item{plotType}{"wc" or "network"}

\item{colorText}{color text label based on frequency in correlation network}

\item{corThresh}{the correlation threshold}

\item{layout}{the layout for correlation network, defaul to "nicely"}

\item{edgeLink}{if FALSE, use geom_edge_diagonal}

\item{deleteZeroDeg}{delete zero degree node from plot in correlation network}

\item{enrich}{currently, only 'reactome' and 'kegg' is supported}

\item{topPath}{how many pathway descriptions are included in text analysis}

\item{ora}{perform ora or not (experimental)}

\item{...}{parameters to pass to wordcloud()}
}
\value{
list of data frame and ggplot2 object
}
\description{
Plot wordcloud of RefSeq description obtained by GeneSummary
}
\examples{
geneList <- c("DDX41")
wcGeneSummary(geneList)
}
